Approximate expected value of decision given perfect information

Calculate the expected value of the conservation management decision given perfect information. This metric describes the value of the management decision that is expected when the decision maker is omniscient (i.e. they know exactly which features occur in which sites). Although real-world conservation planing scenarios rarely involve such omniscient decision makers, this metric is useful to provide an upper bound on the expected value of management decisions following additional data collection.

approx_evdpi(
  site_data,
  feature_data,
  site_occupancy_columns,
  site_probability_columns,
  site_management_cost_column,
  feature_survey_sensitivity_column,
  feature_survey_specificity_column,
  feature_model_sensitivity_column,
  feature_model_specificity_column,
  feature_preweight_column,
  feature_postweight_column,
  feature_target_column,
  total_budget,
  site_management_locked_in_column = NULL,
  prior_matrix = NULL,
  n_approx_obj_fun_points = 1000,
  optimality_gap = 0,
  n_approx_replicates = 100,
  n_approx_states_per_replicate = 1000,
  method_approx_states = "weighted_without_replacement",
  seed = 500
)

Arguments

site_data
sf object with site data.

feature_data
data.frame object with feature data.

site_occupancy_columns
character names of numeric columns in the argument to site_data that contain presence/absence data. Each column should correspond to a different feature, and contain binary presence/absence data (zeros or ones) indicating if the feature was detected in a previous survey or not. If a site has not been surveyed before, then missing (NA) values should be used.

site_probability_columns
character names of numeric columns in the argument to site_data that contain modelled probabilities of occupancy for each feature in each site. Each column should correspond to a different feature, and contain probability data (values between zero and one). No missing (NA) values are permitted in these columns.

site_management_cost_column
character name of column in the argument to site_data that contains costs for managing each site for conservation. This column should have numeric values that are equal to or greater than zero. No missing (NA) values are permitted in this column.

feature_survey_sensitivity_column
character name of the column in the argument to feature_data that contains probability of future surveys correctly detecting a presence of each feature in a given site (i.e. the sensitivity of the survey methodology). This column should have numeric values that are between zero and one. No missing (NA) values are permitted in this column.

feature_survey_specificity_column
character name of the column in the argument to feature_data that contains probability of future surveys correctly detecting an absence of each feature in a given site (i.e. the specificity of the survey methodology). This column should have numeric values that are between zero and one. No missing (NA) values are permitted in this column.

feature_model_sensitivity_column
character name of the column in the argument to feature_data that contains probability of the initial models correctly predicting a presence of each feature in a given site (i.e. the sensitivity of the models). This column should have numeric values that are between zero and one. No missing (NA) values are permitted in this column. This should ideally be calculated using fit_occupancy_models.

feature_model_specificity_column
character name of the column in the argument to feature_data that contains probability of the initial models correctly predicting an absence of each feature in a given site (i.e. the specificity of the models). This column should have numeric values that are between zero and one. No missing (NA) values are permitted in this column. This should ideally be calculated using fit_occupancy_models.

feature_preweight_column
character name of the column in the argument to feature_data that contains the \(preweight\) values used to parametrize the conservation benefit of managing of each feature. This column should have numeric values that are equal to or greater than zero. No missing (NA) values are permitted in this column.

feature_postweight_column
character name of the column in the argument to feature_data that contains the \(postweight\) values used to parametrize the conservation benefit of managing of each feature. This column should have numeric values that are equal to or greater than zero. No missing (NA) values are permitted in this column.

feature_target_column
character name of the column in the argument to feature_data that contains the \(target\) values used to parametrize the conservation benefit of managing of each feature. This column should have numeric values that are equal to or greater than zero. No missing (NA) values are permitted in this column.

total_budget
numeric maximum expenditure permitted for conducting surveys and managing sites for conservation.

site_management_locked_in_column
character name of the column in the argument to site_data that contains logical (TRUE / FALSE) values indicating which sites should be locked in for (TRUE) being managed for conservation or (FALSE) not. No missing (NA) values are permitted in this column. This is useful if some sites have already been earmarked for conservation, or if some sites are already being managed for conservation. Defaults to NULL such that no sites are locked in.

prior_matrix
numeric matrix containing the prior probability of each feature occupying each site. Rows correspond to features, and columns correspond to sites. Defaults to NULL such that prior data is calculated automatically using prior_probability_matrix.

n_approx_obj_fun_points
integer number of points to use for approximating the piecewise-linear components of the objective function. Greater values result in more precise calculations, but also incur more computational costs. Defaults to 1000.

optimality_gap
numeric relative optimality gap for generating conservation prioritizations. A value of zero indicates that prioritizations must be solved to optimality. A value of 0.1 indicates prioritizations must be within 10% of optimality. Defaults to 0.

n_approx_replicates
integer number of replicates to use for approximating the expected value calculations. Defaults to 100.

n_approx_states_per_replicate
integer number of states to use per replicate for approximating the expected value of a given management action. This number must be smaller than or equal to the total number of presence absence states in the system (i.e. n_states(nrow(site_data), nrow(feature_data))) Defaults to 1000.

method_approx_states
character name of method that is used to sample states for approximating the expected value calculations. Available options are: "uniform_with_replacement", "uniform_without_replacement", "weighted_with_replacement", "weighted_without_replacement". Uniform sampling methods have an equal chance of returning each state, and weighted sampling methods are more likely to return states with a higher prior probability of occurring. Defaults to "weighted_without_replacement".

seed
integer random seed used for generating states. Defaults to 500.

Value

numeric vector containing the estimate for each replicate.

See also

Examples

# set seeds for reproducibility
library(RandomFields)
set.seed(123)
RFoptions(seed = 123)

# simulate data
site_data <- simulate_site_data(n_sites = 5, n_features = 2, prop = 0.5)

#> ...

feature_data <- simulate_feature_data(n_features = 2, prop = 1)

# preview simulated data
print(site_data)

#> Simple feature collection with 5 features and 9 fields
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: 0.2875775 ymin: 0.0455565 xmax: 0.9404673 ymax: 0.892419
#> CRS:            NA
#> # A tibble: 5 x 10
#>   survey_cost management_cost     e1     e2     e3    f1    f2    p1     p2
#>         <dbl>           <dbl>  <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl>  <dbl>
#> 1        8.13            73.9 -0.365  0.642  1.08      1     1 0.950 0.940 
#> 2        6.07            69.7 -0.303 -0.382 -0.838     0     0 0.229 0.362 
#> 3        3.74            62.0  1.77   1.42   1.11     NA    NA 0.789 0.0228
#> 4        6.82            71.8 -0.404 -0.771 -0.732    NA    NA 0.157 0.552 
#> 5        5.22            70.7 -0.697 -0.910 -0.614     1     1 0.201 0.798 
#> # ... with 1 more variable: geometry <POINT>

print(feature_data)

#> # A tibble: 2 x 9
#>   name  survey survey_sensitiv~ survey_specific~ model_sensitivi~
#>   <chr> <lgl>             <dbl>            <dbl>            <dbl>
#> 1 f1    TRUE              0.955            0.822            0.775
#> 2 f2    TRUE              0.961            0.829            0.723
#> # ... with 4 more variables: model_specificity <dbl>, preweight <dbl>,
#> #   postweight <dbl>, target <dbl>

# set total budget for managing sites for conservation
total_budget <- sum(site_data$management_cost) * 0.5

# calculate expected value of management decision given perfect information
# using approximate method with 100 replicates and 50 states per replicate
ev_prime_certainty <- approx_evdci(
  site_data, feature_data, c("f1", "f2"), c("p1", "p2"),
  "management_cost", "survey_sensitivity", "survey_specificity",
  "model_sensitivity", "model_specificity",
  "preweight", "postweight", "target",
  total_budget, n_approx_replicates = 100,
  n_approx_states_per_replicate = 50)

# print approximate value
print(ev_prime_certainty)

#>   [1] 27.12129 27.32954 27.10584 27.19117 27.11475 27.16994 27.31391 27.25945
#>   [9] 27.29940 27.08765 27.23246 27.18512 27.29413 27.17236 27.25499 27.23243
#>  [17] 27.25668 27.28042 27.14676 27.15993 27.15378 27.36716 27.24074 27.08906
#>  [25] 27.17746 27.15178 27.16794 27.17173 27.20563 27.10364 27.14101 27.18121
#>  [33] 27.32616 27.20130 27.19479 27.11436 27.23441 27.11150 27.15862 27.05870
#>  [41] 27.10881 27.17384 27.19648 27.20024 27.12909 27.11608 27.05618 27.15185
#>  [49] 27.32096 27.21486 27.18305 27.20491 27.24953 27.28721 27.33608 27.13910
#>  [57] 27.28215 27.22441 27.29708 27.16866 27.10612 27.19141 27.18307 27.15024
#>  [65] 27.18757 27.17249 27.30364 27.11853 27.20869 27.08993 27.19002 27.23383
#>  [73] 27.20443 27.30966 27.18180 27.18960 27.17527 27.21709 27.25418 27.18403
#>  [81] 27.21130 27.22439 27.11031 27.19163 27.25566 27.22908 27.13241 27.23596
#>  [89] 27.25943 27.30171 27.19818 27.20040 27.14392 27.13837 27.07351 27.40606
#>  [97] 27.33158 27.13733 27.10680 27.11466