Expected value of the decision given current information

Calculate the expected value of the management decision given current information. This metric describes the value of the management decision that is expected when the decision maker is limited to existing biodiversity data (i.e. survey data and environmental niche models).

evdci(
  site_data,
  feature_data,
  site_detection_columns,
  site_n_surveys_columns,
  site_probability_columns,
  site_management_cost_column,
  feature_survey_sensitivity_column,
  feature_survey_specificity_column,
  feature_model_sensitivity_column,
  feature_model_specificity_column,
  feature_target_column,
  total_budget,
  site_management_locked_in_column = NULL,
  site_management_locked_out_column = NULL,
  prior_matrix = NULL
)

Arguments

site_data
sf::sf() object with site data.

feature_data
base::data.frame() object with feature data.

site_detection_columns
character names of numeric columns in the argument to site_data that contain the proportion of surveys conducted within each site that detected each feature. Each column should correspond to a different feature, and contain a proportion value (between zero and one). If a site has not previously been surveyed, a value of zero should be used.

site_n_surveys_columns
character names of numeric columns in the argument to site_data that contain the total number of surveys conducted for each each feature within each site. Each column should correspond to a different feature, and contain a non-negative integer number (e.g. 0, 1, 2, 3). If a site has not previously been surveyed, a value of zero should be used.

site_probability_columns
character names of numeric columns in the argument to site_data that contain modelled probabilities of occupancy for each feature in each site. Each column should correspond to a different feature, and contain probability data (values between zero and one). No missing (NA) values are permitted in these columns.

site_management_cost_column
character name of column in the argument to site_data that contains costs for managing each site for conservation. This column should have numeric values that are equal to or greater than zero. No missing (NA) values are permitted in this column.

feature_survey_sensitivity_column
character name of the column in the argument to feature_data that contains probability of future surveys correctly detecting a presence of each feature in a given site (i.e. the sensitivity of the survey methodology). This column should have numeric values that are between zero and one. No missing (NA) values are permitted in this column.

feature_survey_specificity_column
character name of the column in the argument to feature_data that contains probability of future surveys correctly detecting an absence of each feature in a given site (i.e. the specificity of the survey methodology). This column should have numeric values that are between zero and one. No missing (NA) values are permitted in this column.

feature_model_sensitivity_column
character name of the column in the argument to feature_data that contains probability of the initial models correctly predicting a presence of each feature in a given site (i.e. the sensitivity of the models). This column should have numeric values that are between zero and one. No missing (NA) values are permitted in this column. This should ideally be calculated using fit_xgb_occupancy_models() or fit_hglm_occupancy_models().

feature_model_specificity_column
character name of the column in the argument to feature_data that contains probability of the initial models correctly predicting an absence of each feature in a given site (i.e. the specificity of the models). This column should have numeric values that are between zero and one. No missing (NA) values are permitted in this column. This should ideally be calculated using fit_xgb_occupancy_models() or fit_hglm_occupancy_models().

feature_target_column
character name of the column in the argument to feature_data that contains the \(target\) values used to parametrize the conservation benefit of managing of each feature. This column should have numeric values that are equal to or greater than zero. No missing (NA) values are permitted in this column.

total_budget
numeric maximum expenditure permitted for conducting surveys and managing sites for conservation.

site_management_locked_in_column
character name of the column in the argument to site_data that contains logical (TRUE / FALSE) values indicating which sites should be locked in for (TRUE) being managed for conservation or (FALSE) not. No missing (NA) values are permitted in this column. This is useful if some sites have already been earmarked for conservation, or if some sites are already being managed for conservation. Defaults to NULL such that no sites are locked in.

site_management_locked_out_column
character name of the column in the argument to site_data that contains logical (TRUE / FALSE) values indicating which sites should be locked out for (TRUE) being managed for conservation or (FALSE) not. No missing (NA) values are permitted in this column. This is useful if some sites could potentially be surveyed to improve model predictions even if they cannot be managed for conservation. Defaults to NULL such that no sites are locked out.

prior_matrix
numeric matrix containing the prior probability of each feature occupying each site. Rows correspond to features, and columns correspond to sites. Defaults to NULL such that prior data is calculated automatically using prior_probability_matrix().

Value

numeric value.

Details

This function calculates the expected value and does not use approximation methods. As such, this function can only be applied to very small problems.

See also

Examples

# \dontrun{
# set seeds for reproducibility
library(RandomFields)
set.seed(123)
RFoptions(seed = 123)

# simulate data
site_data <- simulate_site_data(n_sites = 5, n_features = 2, prop = 0.5)

#> ...

feature_data <- simulate_feature_data(n_features = 2, prop = 1)
feature_data$target <- c(1, 1)

# preview simulated data
print(site_data)

#> Simple feature collection with 5 features and 11 fields
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: 0.2875775 ymin: 0.0455565 xmax: 0.9404673 ymax: 0.892419
#> CRS:            NA
#> # A tibble: 5 x 12
#>   survey_cost management_cost    f1    f2    n1    n2     e1     e2     e3    p1
#>         <dbl>           <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl>  <dbl>  <dbl> <dbl>
#> 1        8.13            73.9   1       1     5     5 -0.365  0.642  1.08  0.991
#> 2        6.07            69.7   0.2     0     5     5 -0.303 -0.382 -0.838 0.126
#> 3        3.74            62.0   0       0     0     0  1.77   1.42   1.11  0.892
#> 4        6.82            71.8   0       0     2     2 -0.404 -0.771 -0.732 0.064
#> 5        5.22            70.7   0       0     0     0 -0.697 -0.910 -0.614 0.099
#> # … with 2 more variables: p2 <dbl>, geometry <POINT>

print(feature_data)

#> # A tibble: 2 x 7
#>   name  survey survey_sensitiv… survey_specific… model_sensitivi…
#>   <chr> <lgl>             <dbl>            <dbl>            <dbl>
#> 1 f1    TRUE              0.965            0.821            0.701
#> 2 f2    TRUE              0.960            0.814            0.701
#> # … with 2 more variables: model_specificity <dbl>, target <dbl>

# set total budget for managing sites for conservation
total_budget <- sum(site_data$management_cost) * 0.5

# calculate expected value of management decision given current information
# using exact method
ev_current <- evdci(
  site_data, feature_data, c("f1", "f2"), c("n1", "n2"), c("p1", "p2"),
  "management_cost", "survey_sensitivity", "survey_specificity",
  "model_sensitivity", "model_specificity",
  "target", total_budget)

# print exact value
print(ev_current)

#> [1] 1.999758

# }