Expected value of the decision given survey information

Calculate the expected value of the conservation management decision given survey information. This metric describes the value of the management decision that is expected when the decision maker conducts a surveys a set of sites to inform the decision.

evdsi(
  site_data,
  feature_data,
  site_occupancy_columns,
  site_probability_columns,
  site_env_vars_columns,
  site_management_cost_column,
  site_survey_scheme_column,
  site_survey_cost_column,
  feature_survey_column,
  feature_survey_sensitivity_column,
  feature_survey_specificity_column,
  feature_model_sensitivity_column,
  feature_model_specificity_column,
  feature_preweight_column,
  feature_postweight_column,
  feature_target_column,
  total_budget,
  xgb_parameters,
  site_management_locked_in_column = NULL,
  prior_matrix = NULL,
  optimality_gap = 0,
  site_weight_columns = NULL,
  xgb_n_folds = rep(5, nrow(feature_data)),
  seed = 500
)

Arguments

site_data
sf object with site data.

feature_data
data.frame object with feature data.

site_occupancy_columns
character names of numeric columns in the argument to site_data that contain detections and non-detections of features at each site. Each column should correspond to a different feature, and contain binary detection/non-detection value (zeros or ones) indicating if the feature was detected in a previous survey or not. If a site has not been surveyed before, then missing (NA) values should be used. Additionally, if a feature was not looked for when surveying a specific site, then missing (NA) value should be used.

site_probability_columns
character names of numeric columns in the argument to site_data that contain modelled probabilities of occupancy for each feature in each site. Each column should correspond to a different feature, and contain probability data (values between zero and one). No missing (NA) values are permitted in these columns.

site_env_vars_columns
character names of columns in the argument to site_data that contain environmental information for fitting updated occupancy models based on possible survey outcomes. Each column should correspond to a different environmental variable, and contain numeric, factor, or character data. No missing (NA) values are permitted in these columns.

site_management_cost_column
character name of column in the argument to site_data that contains costs for managing each site for conservation. This column should have numeric values that are equal to or greater than zero. No missing (NA) values are permitted in this column.

site_survey_scheme_column
character name of logical (TRUE / FALSE) column in the argument to site_data that indicates which sites are selected in the scheme or not. No missing NA values are permitted. Additionally, only sites that are missing data can be selected or surveying (as per the argument to site_occupancy_columns).

site_survey_cost_column
character name of column in the argument to site_data that contains costs for surveying each site. This column should have numeric values that are equal to or greater than zero. No missing (NA) values are permitted in this column.

feature_survey_column
character name of the column in the argument to feature_data that contains logical (TRUE / FALSE) values indicating if the feature will be surveyed in the planned surveys or not. Note that considering additional features will rapidly increase computational burden, and so it is only recommended to consider features that are of specific conservation interest. No missing (NA) values are permitted in this column.

feature_survey_sensitivity_column
character name of the column in the argument to feature_data that contains probability of future surveys correctly detecting a presence of each feature in a given site (i.e. the sensitivity of the survey methodology). This column should have numeric values that are between zero and one. No missing (NA) values are permitted in this column.

feature_survey_specificity_column
character name of the column in the argument to feature_data that contains probability of future surveys correctly detecting an absence of each feature in a given site (i.e. the specificity of the survey methodology). This column should have numeric values that are between zero and one. No missing (NA) values are permitted in this column.

feature_model_sensitivity_column
character name of the column in the argument to feature_data that contains probability of the initial models correctly predicting a presence of each feature in a given site (i.e. the sensitivity of the models). This column should have numeric values that are between zero and one. No missing (NA) values are permitted in this column. This should ideally be calculated using fit_occupancy_models.

feature_model_specificity_column
character name of the column in the argument to feature_data that contains probability of the initial models correctly predicting an absence of each feature in a given site (i.e. the specificity of the models). This column should have numeric values that are between zero and one. No missing (NA) values are permitted in this column. This should ideally be calculated using fit_occupancy_models.

feature_preweight_column
character name of the column in the argument to feature_data that contains the \(preweight\) values used to parametrize the conservation benefit of managing of each feature. This column should have numeric values that are equal to or greater than zero. No missing (NA) values are permitted in this column.

feature_postweight_column
character name of the column in the argument to feature_data that contains the \(postweight\) values used to parametrize the conservation benefit of managing of each feature. This column should have numeric values that are equal to or greater than zero. No missing (NA) values are permitted in this column.

feature_target_column
character name of the column in the argument to feature_data that contains the \(target\) values used to parametrize the conservation benefit of managing of each feature. This column should have numeric values that are equal to or greater than zero. No missing (NA) values are permitted in this column.

total_budget
numeric maximum expenditure permitted for conducting surveys and managing sites for conservation.

xgb_parameters
list of list objects containing the parameters for fitting models for each feature. See documentation for the params argument in xgb.train for available parameters. Ideally, these parameters would be determined using the fit_occupancy_models function. Note that arguments must have "nrounds", "objective", "scale_pos_weight" elements (see example below).

site_management_locked_in_column
character name of the column in the argument to site_data that contains logical (TRUE / FALSE) values indicating which sites should be locked in for (TRUE) being managed for conservation or (FALSE) not. No missing (NA) values are permitted in this column. This is useful if some sites have already been earmarked for conservation, or if some sites are already being managed for conservation. Defaults to NULL such that no sites are locked in.

prior_matrix
numeric matrix containing the prior probability of each feature occupying each site. Rows correspond to features, and columns correspond to sites. Defaults to NULL such that prior data is calculated automatically using prior_probability_matrix.

optimality_gap
numeric relative optimality gap for generating conservation prioritizations. A value of zero indicates that prioritizations must be solved to optimality. A value of 0.1 indicates prioritizations must be within 10% of optimality. Defaults to 0.

site_weight_columns
character name of columns in site_data containing weights for model fitting for each feature. These columns must contain numeric values greater than or equal to zero. No missing (NA) values are permitted. Defaults to NULL such that all data are given equal weight when fitting models.

xgb_n_folds
integer vector containing the number of k-fold cross-validation folds to use for fitting models and assessing model performance for each feature. Ideally, the number of folds should be exactly the same as the number used for tuning the model parameters (i.e. same parameter to the n_folds argument in fit_occupancy_models when generating parameters for xgb_parameters).

seed
integer state of the random number generator for partitioning data into folds cross-validation and fitting xgboost models. This parameter must remain the same to compare results for different survey schemes. Defaults to 500.

Value

numeric value.

Details

This function calculates the expected value and does not use approximation methods. As such, this function can only be applied to very small problems.

See also

Examples

# set seeds for reproducibility
library(RandomFields)
set.seed(123)
RFoptions(seed = 123)

# simulate data
site_data <- simulate_site_data(n_sites = 15, n_features = 2, prop = 0.5)

#> ...

feature_data <- simulate_feature_data(n_features = 2, prop = 1)

# preview simulated data
print(site_data)

#> Simple feature collection with 15 features and 9 fields
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: 0.0455565 ymin: 0.04205953 xmax: 0.9568333 ymax: 0.9942698
#> CRS:            NA
#> # A tibble: 15 x 10
#>    survey_cost management_cost       e1     e2      e3    f1    f2     p1
#>          <dbl>           <dbl>    <dbl>  <dbl>   <dbl> <dbl> <dbl>  <dbl>
#>  1        5.73            64.1 -0.149    0.803  0.298     NA    NA 0.888 
#>  2        3.50            65.9 -0.00413 -0.437 -1.99       0     0 0.0455
#>  3        6.23            67.0  1.71     1.27  -0.120     NA    NA 0.438 
#>  4        4.24            69.0 -0.412   -1.04  -1.88       0     0 0.0307
#>  5        6.19            67.2 -0.733   -0.930  1.87       0     1 0.803 
#>  6        6.15            61.7  0.956    0.297 -0.288     NA    NA 0.237 
#>  7        7.01            67.0 -0.503    0.229  0.341     NA    NA 0.825 
#>  8        4.45            64.4 -1.00    -1.54   0.147     NA    NA 0.219 
#>  9        6.47            66.6 -1.03     0.707  1.05      NA    NA 0.982 
#> 10        6.92            66.8 -0.152    0.489  0.240      1     0 0.805 
#> 11        4.60            70.8 -0.954   -1.74   0.633      1     1 0.238 
#> 12        6.71            71.4  0.177    0.509  0.0210    NA    NA 0.673 
#> 13        5.96            70.3 -0.714   -0.541 -0.0777     1     0 0.486 
#> 14        3.64            67.7  0.481    0.271 -0.868      0     0 0.238 
#> 15        7.95            72.8  2.33     1.64   0.629     NA    NA 0.595 
#> # ... with 2 more variables: p2 <dbl>, geometry <POINT>

print(feature_data)

#> # A tibble: 2 x 9
#>   name  survey survey_sensitiv~ survey_specific~ model_sensitivi~
#>   <chr> <lgl>             <dbl>            <dbl>            <dbl>
#> 1 f1    TRUE              0.964            0.827            0.703
#> 2 f2    TRUE              0.960            0.835            0.800
#> # ... with 4 more variables: model_specificity <dbl>, preweight <dbl>,
#> #   postweight <dbl>, target <dbl>

# set total budget for managing sites for conservation
 # (i.e. 50% of the cost of managing all sites)
total_budget <- sum(site_data$management_cost) * 0.5

# create a survey scheme that samples the first two sites that
# are missing data
site_data$survey_site <- FALSE
site_data$survey_site[which(is.na(site_data$f1))[1:2]] <- TRUE

# define xgboost tuning parameters
# these should ideally be determined using fit_occupancy_models
xgb_parameters <-
 list(list(objective = "binary:logistic", nrounds = 8,
           scale_pos_weight = 1, eta = 0.1))[rep(1, 2)]

# calculate expected value of management decision given the survey
# information using exact method
ev_survey <- evdsi(
  site_data, feature_data, c("f1", "f2"), c("p1", "p2"),
  c("e1", "e2", "e3"), "management_cost", "survey_site",
  "survey_cost", "survey", "survey_sensitivity", "survey_specificity",
  "model_sensitivity", "model_specificity",
  "preweight", "postweight", "target",
  total_budget, xgb_parameters)

# print exact value
print(ev_survey)

#> [1] 89.51493