Expected value of the decision given current information

Calculate the expected value of the conservation management decision given current information. This metric describes the value of the management decision that is expected when the decision maker is limited to existing biodiversity data (i.e. survey data and environmental niche models).

expected_value_of_decision_given_current_information(site_data,
  feature_data, site_occupancy_columns, site_probability_columns,
  site_management_cost_column, feature_survey_sensitivity_column,
  feature_survey_specificity_column, feature_model_sensitivity_column,
  feature_alpha_column, feature_gamma_column, total_budget,
  site_management_locked_in_column = NULL, prior_matrix = NULL,
  n_approx_obj_fun_points = 1000, optimality_gap = 0)

Arguments

site_data
sf object with site data.

feature_data
data.frame object with feature data.

site_occupancy_columns
character names of numeric columns in the argument to site_data that contain presence/absence data. Each column should correspond to a different feature, and contain binary presence/absence data (zeros or ones) indicating if the feature was detected in a previous survey or not. If a site has not been surveyed before, then missing (NA) values should be used.

site_probability_columns
character names of numeric columns in the argument to site_data that contain modelled probabilities of occupancy for each feature in each site. Each column should correspond to a different feature, and contain probability data (values between zero and one). No missing (NA) values are permitted in these columns.

site_management_cost_column
character name of column in the argument to site_data that contains costs for managing each site for conservation. This column should have numeric values that are equal to or greater than zero. No missing (NA) values are permitted in this column.

feature_survey_sensitivity_column
character name of the column in the argument to feature_data that contains probability of future surveys correctly detecting a presence of each feature in a given site (i.e. the sensitivity of the survey methodology). This column should have numeric values that are between zero and one. No missing (NA) values are permitted in this column.

feature_survey_specificity_column
character name of the column in the argument to feature_data that contains probability of future surveys correctly detecting an absence of each feature in a given site (i.e. the specificity of the survey methodology). This column should have numeric values that are between zero and one. No missing (NA) values are permitted in this column.

feature_model_sensitivity_column
character name of the column in the argument to feature_data that contains probability of the initial models correctly predicting a presence of each feature in a given site (i.e. the sensitivity of the models). This column should have numeric values that are between zero and one. No missing (NA) values are permitted in this column. This should ideally be calculated using fit_occupancy_models.

feature_alpha_column
character name of the column in the argument to feature_data that contains the \(\alpha\) values used to parametrize the conservation benefit of managing of each feature. This column should have numeric values that are equal to or greater than zero. No missing (NA) values are permitted in this column.

feature_gamma_column
character name of the column in the argument to feature_data that contains the \(\gamma\) values used to parametrize the conservation benefit of managing of each feature. This column should have numeric values that are equal to or greater than zero. No missing (NA) values are permitted in this column.

total_budget
numeric maximum expenditure permitted for conducting surveys and managing sites for conservation.

site_management_locked_in_column
character name of the column in the argument to site_data that contains logical (TRUE / FALSE) values indicating which sites should be locked in for (TRUE) being managed for conservation or (FALSE) not. No missing (NA) values are permitted in this column. This is useful if some sites have already been earmarked for conservation, or if some sites are already being managed for conservation. Defaults to NULL such that no sites are locked in.

prior_matrix
numeric matrix containing the prior probability of each feature occupying each site. Rows correspond to features, and columns correspond to sites. Defaults to NULL such that prior data is calculated automatically using prior_probability_matrix.

n_approx_obj_fun_points
integer number of points to use for approximating the piecewise-linear components of the objective function. Greater values result in more precise calculations, but also incur more computational costs. Defaults to 1000.

optimality_gap
numeric relative optimality gap for generating conservation prioritizations. A value of zero indicates that prioritizations must be solved to optimality. A value of 0.1 indicates prioritizations must be within 10% of optimality. Defaults to 0.

Value

numeric value.

Details

Let \(I\) denote the set of feature (indexed by \(i\)) and \(J\) the set of planning units (indexed by \(j\)). To describe the results of previous surveys, let \(D_{ij}\) indicate the detection/non-detection of feature \(i \in I\) in planning unit \(j \in J\) during previous surveys using zeros and ones (specified via site_occupancy_columns). Also let \(U_j\) indicate if planning unit \(j \in J\) has previously been surveyed or not. To describe the accuracy of the surveys, let \(S_i\) denote the sensitivity and \(N_i\) denote the specificity of the surveying methodology for features \(i \in I\) (specified via feature_survey_sensitivity_column and feature_survey_specificity_column respectively). Next, let \(H_{ij}\) denote the modelled probabilities of the features \(i \in I\) occupying planning units \(j \in J\) (specified via site_probability_columns). To describe the accuracy of the environmental niche models, let \({S'}_i\) denote the sensitivity of the models for features \(i \in I\) (specified via feature_model_sensitivity_column). We can calculate the prior probability of each feature \(i \in I\) occupying each site \(j \in J\) following (or manually specified via prior_matrix):

$$ Q_{ij} = \\ S_i, \text{ if } D_j = 1, H_{ij} = 1 \space (i \text{ detected in } j) \\ 1 - N_i, \text{ else if } D_j = 1, H_{ij} = 0 \space (i \text{ not detected in } j) \\ {S'}_i {H'}_{ij}, \text{ else if } D_j = 0 \space (j \text{ not surveyed)} \\ $$

Since we do not know which features truly occur in which sites, let \(S\) denote the full range of possible states that affect the evaluation of the management action (indexed by \(s\)). Here, \(S\) represents all possible combinations of different species occurring in different sites. The total number of states \(s \in S\) is incredibly large for even a small conservation planning problem. For example, a conservation planning problem involving ten sites and four features (species) has approximately \(1.09 \times 10^{12}\) states. In other words, one trillion ninety-nine billion five hundred eleven million six hundred twenty-seven thousand seven hundred seventy-five states. Unfortunately, this means that it is not computationally feasible to calculate exact values of information for conservation problems involving many sites. Let \(G_{ijs}\) indicate which species \(i \in I\) occur in which planning units \(j \in J\) given state \(s \in S\). We calculate the prior probability of each state \(s \in S\) being the true state following:

$$P_s = \\ Q_{ij}, \text{ if } G_{ijs} = 1 \\ 1 - Q_{ij}, \text{ else } \\ $$

Here, the management action is to purchase a set of sites for protected area establishment. In other words, the management action is to implement a spatial prioritisation for protected area establishment. The management action is subject to a budget that limits the cost of purchasing sites. Let \(A_j\) denote the cost of acquiring each site \(j \in J\) (specified via site_management_cost_column), and \(b\) denote the total budget available for conservation (specified via total_budget). Furthermore, let \(T_j\) denote which sites \(j \in j\) are locked into the prioritization (e.g. sites in existing protected area system). Given these data, let \(Z\) denote the set of all possible management actions (indexed by \(z\)). To describe each of the possible management actions \(z \in Z\), let \(X_{jz}\) indicate which sites \(j \in J\) are purchased for protection (using zeros and ones) under a given management action \(z \in Z\). To evaluate the conservation value of a given management action, we will adopt equations used in the objective function of the Zonation decision support tool (Moilanen 2007). Let \(\alpha_i\) and \(\gamma_i\) denote scaling terms for species \(i \in I\) to parametrise trade-offs in the protection of different species. These scaling terms are crucial to ensure that competing management decisions are evaluated in a manner conforming to the principle of complementarity (Vane-Wright et al. 1991). The value of a management action (\(z\)) for a given state (\(s\)) is calculated following, wherein \(Y_j\) indicates if each site \(j \in J\) is selected in the (\(z\)) management action or not (using zeros and ones):

$$ V(z, s) = \sum_{i \in I} \alpha_i \left( \sum_{j \in J} Y_{j} G_{ijs} \right) ^{\gamma_i} \\ $$

We can now calculate the expected value of the management decision given current information (\(\text{EV}_{\text{current}}\)). To achieve this, we assume that the decision maker will act optimally. Thus the expected value of the management decision given current information is the expected value of the best management action given current information. Let \(z'\) denote an optimal management action \(z \in Z\) given current information. This can be expressed as follows:

$$ \text{EV}_{\text{current}} = \mathbb{E} \left[ V(z', S) \right] = \max_{z \in Z} \mathbb{E} \left[ V(z, S) \right] \\ \mathbb{E} \left[ V(z, S) \right] = \sum_{s \in S} V(z, s) \times P_s \\ $$

Typically, the emphexpected value of the best management action given current information is calculated by iterating over every possible management action \(z \in Z\). However, this approach would be computationally expensive for even a small conservation planning problem. For example, a conservation planning problem involving 60 planning units has approximately \(1.52 \times 10^{18}\) different solutions (i.e. combinations of different planning units being selected). Therefore, we applied combinatorial optimization instead. Specifically, we identify the best management action by formulating an integer programming problem -- using piecewise linear constraints to linearise the objective function (precision controlled using n_approx_n_approx_obj_fun_points) -- and solving it to (near) optimality (controlled using optimality_gap) with the gurobi package. Let \({X'}_j\) indicate if each planning unit \(j \in J\) is selected in the \(z'\) best management action given current information (using zeros and ones):

$$ \text{maximize } \sum_{i \in I} \alpha_i \left( \sum_{j \in J} {X'}_j Q_{ij} \right) ^{\gamma_i} \\ \text{subject to } \sum_{j \in J} {{X'}_j} A_j \leq b \\ {{X'}_j} \geq T_j \text{ } \forall j \in J \\ {{X'}_j} \in \{ 0, 1 \} \text{ } \forall j \in J $$

References

Moilanen A (2007) Landscape zonation, benefit functions and target-based planning: unifying reserve selection strategies. Biological Conservation, 134: 571--579.

Vane-Wright RI, Humphries CJ, & Williams PH (1991) What to protect? Systematics and the agony of choice. Biological Conservation, 55: 235--254.

See also

Examples

# set seeds for reproducibility
library(RandomFields)
set.seed(123)
RFoptions(seed = 123)

# simulate data
site_data <- simulate_site_data(n_sites = 5, n_features = 2, prop = 0.5)

#> ...

feature_data <- simulate_feature_data(n_features = 2, prop = 1)

# preview simulated data
print(site_data)

#> Simple feature collection with 5 features and 9 fields
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: 0.2875775 ymin: 0.0455565 xmax: 0.9404673 ymax: 0.892419
#> epsg (SRID):    NA
#> proj4string:    NA
#> # A tibble: 5 x 10
#>   survey_cost management_cost     e1     e2     e3    f1    f2    p1     p2
#>         <dbl>           <dbl>  <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl>  <dbl>
#> 1        8.13            73.9 -0.365  0.642  1.08      1     1 0.950 0.940 
#> 2        6.07            69.7 -0.303 -0.382 -0.838     0     0 0.229 0.362 
#> 3        3.74            62.0  1.77   1.42   1.11     NA    NA 0.789 0.0228
#> 4        6.82            71.8 -0.404 -0.771 -0.732    NA    NA 0.157 0.552 
#> 5        5.22            70.7 -0.697 -0.910 -0.614     1     1 0.201 0.798 
#> # ... with 1 more variable: geometry <POINT>

print(feature_data)

#> # A tibble: 2 x 7
#>   name  survey survey_sensitivi~ survey_specifici~ model_sensitivi~ alpha  gamma
#>   <chr> <lgl>              <dbl>             <dbl>            <dbl> <dbl>  <dbl>
#> 1 f1    TRUE               0.955             0.822            0.775  2.70 0.642 
#> 2 f2    TRUE               0.961             0.829            0.723  2.95 0.0777

# set total budget for managing sites for conservation
total_budget <- sum(site_data$management_cost) * 0.5

# calculate expected value of management decision given current information
# using exact method
ev_current <- expected_value_of_decision_given_current_information(
  site_data, feature_data, c("f1", "f2"), c("p1", "p2"),
  "management_cost", "survey_sensitivity",
  "survey_specificity", "model_sensitivity", "alpha", "gamma", total_budget)

# print exact value
print(ev_current)

#> [1] 3.995959