Fit hierarchical generalized linear models to predict occupancy

Estimate probability of occupancy for a set of features in a set of planning units. Models are fitted as hierarchical generalized linear models that account for for imperfect detection (following Royle & Link 2006) using JAGS (via run.jags). To limit over-fitting, spike-and-slab coefficients are used to perform stochastic search variable selection (Kuo & Mallick 1998).

fit_hglm_occupancy_models(
  site_data,
  feature_data,
  site_detection_columns,
  site_n_surveys_columns,
  site_env_vars_columns,
  feature_survey_sensitivity_column,
  feature_survey_specificity_column,
  jags_n_samples = rep(10000, length(site_detection_columns)),
  jags_n_burnin = rep(1000, length(site_detection_columns)),
  jags_n_thin = rep(100, length(site_detection_columns)),
  jags_n_adapt = rep(1000, length(site_detection_columns)),
  jags_n_chains = rep(4, length(site_detection_columns)),
  n_folds = rep(5, length(site_detection_columns)),
  n_threads = 1,
  seed = 500,
  verbose = FALSE
)

Arguments

site_data
sf object with site data.

feature_data
data.frame object with feature data.

site_detection_columns
character names of numeric columns in the argument to site_data that contain the proportion of surveys conducted within each site that detected each feature. Each column should correspond to a different feature, and contain a proportion value (between zero and one). If a site has not previously been surveyed, a value of zero should be used.

site_n_surveys_columns
character names of numeric columns in the argument to site_data that contain the total number of surveys conducted for each each feature within each site. Each column should correspond to a different feature, and contain a non-negative integer number (e.g. 0, 1, 2, 3). If a site has not previously been surveyed, a value of zero should be used.

site_env_vars_columns
character names of columns in the argument to site_data that contain environmental information for fitting updated occupancy models based on possible survey outcomes. Each column should correspond to a different environmental variable, and contain numeric, factor, or character data. No missing (NA) values are permitted in these columns.

feature_survey_sensitivity_column
character name of the column in the argument to feature_data that contains probability of future surveys correctly detecting a presence of each feature in a given site (i.e. the sensitivity of the survey methodology). This column should have numeric values that are between zero and one. No missing (NA) values are permitted in this column.

feature_survey_specificity_column
character name of the column in the argument to feature_data that contains probability of future surveys correctly detecting an absence of each feature in a given site (i.e. the specificity of the survey methodology). This column should have numeric values that are between zero and one. No missing (NA) values are permitted in this column.

n_folds
numeric number of folds to split the training data into when fitting models for each feature. Defaults to 5 for each feature.

n_threads
integer number of threads to use for parameter tuning. Defaults to 1.

seed
integer initial random number generator state for model fitting. Defaults to 500.

verbose
logical indicating if information should be printed during computations. Defaults to FALSE.

Value

list object containing:

models

list of list objects containing the models.

predictions

tibble object containing predictions for each feature.

performance

tibble object containing the performance of the best models for each feature. It contains the following columns:
feature

name of the feature.

max_mpsrf

maximum multi-variate potential scale reduction factor (MPSRF) value for the models. A MPSRF value less than 1.05 means that all coefficients in a given model have converged, and so a value less than 1.05 in this column means that all the models fit for a given feature have successfully converged.

train_tss_mean

mean TSS statistic for models calculated using training data in cross-validation.

train_tss_std

standard deviation in TSS statistics for models calculated using training data in cross-validation.

train_sensitivity_mean

mean sensitivity statistic for models calculated using training data in cross-validation.

train_sensitivity_std

standard deviation in sensitivity statistics for models calculated using training data in cross-validation.

train_specificity_mean

mean specificity statistic for models calculated using training data in cross-validation.

train_specificity_std

standard deviation in specificity statistics for models calculated using training data in cross-validation.

test_tss_mean

mean TSS statistic for models calculated using test data in cross-validation.

test_tss_std

standard deviation in TSS statistics for models calculated using test data in cross-validation.

test_sensitivity_mean

mean sensitivity statistic for models calculated using test data in cross-validation.

test_sensitivity_std

standard deviation in sensitivity statistics for models calculated using test data in cross-validation.

test_specificity_mean

mean specificity statistic for models calculated using test data in cross-validation.

test_specificity_std

standard deviation in specificity statistics for models calculated using test data in cross-validation.

Details

This function (i) prepares the data for model fitting, (ii) fits the models, and (iii) assesses the performance of the models. These analyses are performed separately for each feature. For a given feature:

The data are prepared for model fitting by partitioning the data using k-fold cross-validation (set via argument to n_folds). The training and evaluation folds are constructed in such a manner as to ensure that each training and evaluation fold contains at least one presence and one absence observation.

A model for fit separately for each fold (see inst/jags/model.jags for model code). To assess convergence, the multi-variate potential scale reduction factor (MPSRF) statistic is calculated for each model.

The performance of the cross-validation models is evaluated. Specifically, the TSS, sensitivity, and specificity statistics are calculated (if relevant, weighted by the argument to site_weights_data). These performance values are calculated using the models' training and evaluation folds. To assess convergence, the maximum MPSRF statistic for the models fit for each feature is calculated.

References

Kuo L & Mallick B (1998) Variable selection for regression models. Sankhyā: The Indian Journal of Statistics, Series B, 60: 65--81.

Royle JA & Link WA (2006) Generalized site occupancy models allowing for false positive and false negative errors. Ecology, 87: 835--841.

Examples

# set seeds for reproducibility
library(RandomFields)

#> Loading required package: sp

#> Loading required package: RandomFieldsUtils

#> 
#> Attaching package: ‘RandomFields’

#> The following object is masked from ‘package:RandomFieldsUtils’:
#> 
#>     RFoptions

set.seed(123)
RFoptions(seed = 123)

# simulate data for 200 sites, 2 features, and 3 environmental variables
site_data <- simulate_site_data(n_sites = 30, n_features = 2, prop = 0.1)

#> ...

feature_data <- simulate_feature_data(n_features = 2, prop = 1)

# print JAGS model code
cat(readLines(system.file("jags", "model.jags", package = "surveyvoi")),
    sep = "\n")

#> model {
#>   # data
#>   ## int<lower=1> n_vars
#>   ## int<lower=1> train_n_sites
#>   ## real sens
#>   ## real spec
#>   ## real[train_n_sites, n_vars] train_model_matrix
#>   ## real[train_n_sites] train_n_obs
#>   ## matrix[train_n_sites, max(train_n_obs)] train_obs
#> 
#>   # priors
#>   for (i in 1:n_vars) {
#>     coefs_raw[i] ~ dnorm(0, 2.5) # raw model coefficients
#>     ss[i] ~ dbern(0.5)           # spike-and-slab variable selection terms
#>   }
#> 
#>   # transformed parameters
#>   for (i in 1:n_vars) {
#>     coefs[i] <- coefs_raw[i] * ss[i] # raw model coefficients
#>   }
#>   for (i in 1:train_n_sites) {
#>     logit(yhat[i]) <- coefs %*% train_model_matrix[i, ]
#>   }
#> 
#>   # likelihood
#>   for (i in 1:train_n_sites) {
#>     occ[i] ~ dbern(yhat[i])
#>     for (j in 1:train_n_obs[i]) {
#>       train_obs[i, j] ~ dbern((occ[i] * sens) + ((1 - occ[i]) * spec))
#>     }
#>   }
#> }

# \dontrun{
# fit models
# note that we use a small number of MCMC iterations so that the example
# finishes quickly, you probably want to use the defaults for real work
results <- fit_hglm_occupancy_models(
   site_data, feature_data,
   c("f1", "f2"), c("n1", "n2"), c("e1", "e2", "e3"),
   "survey_sensitivity", "survey_specificity",
   n_folds = rep(5, 2),
   jags_n_samples = rep(250, 2), jags_n_burnin = rep(250, 2),
   jags_n_thin = rep(1, 2), jags_n_adapt = rep(100, 2),
   n_threads = 1)

#>   |                                                                              |                                                                      |   0%

#> Loading required namespace: rjags

#> Compiling rjags model...
#> Calling the simulation using the rjags method...
#> Adapting the model for 100 iterations...
#> Burning in the model for 250 iterations...
#> Running the model for 250 iterations...
#> Simulation complete
#> Calculating summary statistics...
#> Finished running the simulation
#>   |                                                                              |==                                                                    |   2%Compiling rjags model...
#> Calling the simulation using the rjags method...
#> Adapting the model for 100 iterations...
#> Burning in the model for 250 iterations...
#> Running the model for 250 iterations...
#> Simulation complete
#> Calculating summary statistics...
#> Finished running the simulation
#>   |                                                                              |====                                                                  |   5%Compiling rjags model...
#> Calling the simulation using the rjags method...
#> Adapting the model for 100 iterations...
#> Burning in the model for 250 iterations...
#> Running the model for 250 iterations...
#> Simulation complete
#> Calculating summary statistics...
#> Finished running the simulation
#>   |                                                                              |=====                                                                 |   8%Compiling rjags model...
#> Calling the simulation using the rjags method...
#> Adapting the model for 100 iterations...
#> Burning in the model for 250 iterations...
#> Running the model for 250 iterations...
#> Simulation complete
#> Calculating summary statistics...
#> Finished running the simulation
#>   |                                                                              |=======                                                               |  10%Compiling rjags model...
#> Calling the simulation using the rjags method...
#> Adapting the model for 100 iterations...
#> Burning in the model for 250 iterations...
#> Running the model for 250 iterations...
#> Simulation complete
#> Calculating summary statistics...
#> Finished running the simulation
#>   |                                                                              |=========                                                             |  12%Compiling rjags model...
#> Calling the simulation using the rjags method...
#> Adapting the model for 100 iterations...
#> Burning in the model for 250 iterations...
#> Running the model for 250 iterations...
#> Simulation complete
#> Calculating summary statistics...
#> Finished running the simulation
#>   |                                                                              |==========                                                            |  15%Compiling rjags model...
#> Calling the simulation using the rjags method...
#> Adapting the model for 100 iterations...
#> Burning in the model for 250 iterations...
#> Running the model for 250 iterations...
#> Simulation complete
#> Calculating summary statistics...
#> Finished running the simulation
#>   |                                                                              |============                                                          |  18%Compiling rjags model...
#> Calling the simulation using the rjags method...
#> Adapting the model for 100 iterations...
#> Burning in the model for 250 iterations...
#> Running the model for 250 iterations...
#> Simulation complete
#> Calculating summary statistics...
#> Finished running the simulation
#>   |                                                                              |==============                                                        |  20%Compiling rjags model...
#> Calling the simulation using the rjags method...
#> Adapting the model for 100 iterations...
#> Burning in the model for 250 iterations...
#> Running the model for 250 iterations...
#> Simulation complete
#> Calculating summary statistics...
#> Finished running the simulation
#>   |                                                                              |================                                                      |  22%Compiling rjags model...
#> Calling the simulation using the rjags method...
#> Adapting the model for 100 iterations...
#> Burning in the model for 250 iterations...
#> Running the model for 250 iterations...
#> Simulation complete
#> Calculating summary statistics...
#> Finished running the simulation
#>   |                                                                              |==================                                                    |  25%Compiling rjags model...
#> Calling the simulation using the rjags method...
#> Adapting the model for 100 iterations...
#> Burning in the model for 250 iterations...
#> Running the model for 250 iterations...
#> Simulation complete
#> Calculating summary statistics...
#> Finished running the simulation
#>   |                                                                              |===================                                                   |  28%Compiling rjags model...
#> Calling the simulation using the rjags method...
#> Adapting the model for 100 iterations...
#> Burning in the model for 250 iterations...
#> Running the model for 250 iterations...
#> Simulation complete
#> Calculating summary statistics...
#> Finished running the simulation
#>   |                                                                              |=====================                                                 |  30%Compiling rjags model...
#> Calling the simulation using the rjags method...
#> Adapting the model for 100 iterations...
#> Burning in the model for 250 iterations...
#> Running the model for 250 iterations...
#> Simulation complete
#> Calculating summary statistics...
#> Finished running the simulation
#>   |                                                                              |=======================                                               |  32%Compiling rjags model...
#> Calling the simulation using the rjags method...
#> Adapting the model for 100 iterations...
#> Burning in the model for 250 iterations...
#> Running the model for 250 iterations...
#> Simulation complete
#> Calculating summary statistics...
#> Finished running the simulation
#>   |                                                                              |========================                                              |  35%Compiling rjags model...
#> Calling the simulation using the rjags method...
#> Adapting the model for 100 iterations...
#> Burning in the model for 250 iterations...
#> Running the model for 250 iterations...
#> Simulation complete
#> Calculating summary statistics...
#> Finished running the simulation
#>   |                                                                              |==========================                                            |  38%Compiling rjags model...
#> Calling the simulation using the rjags method...
#> Adapting the model for 100 iterations...
#> Burning in the model for 250 iterations...
#> Running the model for 250 iterations...
#> Simulation complete
#> Calculating summary statistics...
#> Finished running the simulation
#>   |                                                                              |============================                                          |  40%Compiling rjags model...
#> Calling the simulation using the rjags method...
#> Adapting the model for 100 iterations...
#> Burning in the model for 250 iterations...
#> Running the model for 250 iterations...
#> Simulation complete
#> Calculating summary statistics...
#> Finished running the simulation
#>   |                                                                              |==============================                                        |  42%Compiling rjags model...
#> Calling the simulation using the rjags method...
#> Adapting the model for 100 iterations...
#> Burning in the model for 250 iterations...
#> Running the model for 250 iterations...
#> Simulation complete
#> Calculating summary statistics...
#> Finished running the simulation
#>   |                                                                              |================================                                      |  45%Compiling rjags model...
#> Calling the simulation using the rjags method...
#> Adapting the model for 100 iterations...
#> Burning in the model for 250 iterations...
#> Running the model for 250 iterations...
#> Simulation complete
#> Calculating summary statistics...
#> Finished running the simulation
#>   |                                                                              |=================================                                     |  48%Compiling rjags model...
#> Calling the simulation using the rjags method...
#> Adapting the model for 100 iterations...
#> Burning in the model for 250 iterations...
#> Running the model for 250 iterations...
#> Simulation complete
#> Calculating summary statistics...
#> Finished running the simulation
#>   |                                                                              |===================================                                   |  50%Compiling rjags model...
#> Calling the simulation using the rjags method...
#> Adapting the model for 100 iterations...
#> Burning in the model for 250 iterations...
#> Running the model for 250 iterations...
#> Simulation complete
#> Calculating summary statistics...
#> Finished running the simulation
#>   |                                                                              |=====================================                                 |  52%Compiling rjags model...
#> Calling the simulation using the rjags method...
#> Adapting the model for 100 iterations...
#> Burning in the model for 250 iterations...
#> Running the model for 250 iterations...
#> Simulation complete
#> Calculating summary statistics...
#> Finished running the simulation
#>   |                                                                              |======================================                                |  55%Compiling rjags model...
#> Calling the simulation using the rjags method...
#> Adapting the model for 100 iterations...
#> Burning in the model for 250 iterations...
#> Running the model for 250 iterations...
#> Simulation complete
#> Calculating summary statistics...
#> Finished running the simulation
#>   |                                                                              |========================================                              |  58%Compiling rjags model...
#> Calling the simulation using the rjags method...
#> Adapting the model for 100 iterations...
#> Burning in the model for 250 iterations...
#> Running the model for 250 iterations...
#> Simulation complete
#> Calculating summary statistics...
#> Finished running the simulation
#>   |                                                                              |==========================================                            |  60%Compiling rjags model...
#> Calling the simulation using the rjags method...
#> Adapting the model for 100 iterations...
#> Burning in the model for 250 iterations...
#> Running the model for 250 iterations...
#> Simulation complete
#> Calculating summary statistics...
#> Finished running the simulation
#>   |                                                                              |============================================                          |  62%Compiling rjags model...
#> Calling the simulation using the rjags method...
#> Adapting the model for 100 iterations...
#> Burning in the model for 250 iterations...
#> Running the model for 250 iterations...
#> Simulation complete
#> Calculating summary statistics...
#> Finished running the simulation
#>   |                                                                              |==============================================                        |  65%Compiling rjags model...
#> Calling the simulation using the rjags method...
#> Adapting the model for 100 iterations...
#> Burning in the model for 250 iterations...
#> Running the model for 250 iterations...
#> Simulation complete
#> Calculating summary statistics...
#> Finished running the simulation
#>   |                                                                              |===============================================                       |  68%Compiling rjags model...
#> Calling the simulation using the rjags method...
#> Adapting the model for 100 iterations...
#> Burning in the model for 250 iterations...
#> Running the model for 250 iterations...
#> Simulation complete
#> Calculating summary statistics...
#> Finished running the simulation
#>   |                                                                              |=================================================                     |  70%Compiling rjags model...
#> Calling the simulation using the rjags method...
#> Adapting the model for 100 iterations...
#> Burning in the model for 250 iterations...
#> Running the model for 250 iterations...
#> Simulation complete
#> Calculating summary statistics...
#> Finished running the simulation
#>   |                                                                              |===================================================                   |  72%Compiling rjags model...
#> Calling the simulation using the rjags method...
#> Adapting the model for 100 iterations...
#> Burning in the model for 250 iterations...
#> Running the model for 250 iterations...
#> Simulation complete
#> Calculating summary statistics...
#> Finished running the simulation
#>   |                                                                              |====================================================                  |  75%Compiling rjags model...
#> Calling the simulation using the rjags method...
#> Adapting the model for 100 iterations...
#> Burning in the model for 250 iterations...
#> Running the model for 250 iterations...
#> Simulation complete
#> Calculating summary statistics...
#> Finished running the simulation
#>   |                                                                              |======================================================                |  78%Compiling rjags model...
#> Calling the simulation using the rjags method...
#> Adapting the model for 100 iterations...
#> Burning in the model for 250 iterations...
#> Running the model for 250 iterations...
#> Simulation complete
#> Calculating summary statistics...
#> Finished running the simulation
#>   |                                                                              |========================================================              |  80%Compiling rjags model...
#> Calling the simulation using the rjags method...
#> Adapting the model for 100 iterations...
#> Burning in the model for 250 iterations...
#> Running the model for 250 iterations...
#> Simulation complete
#> Calculating summary statistics...
#> Finished running the simulation
#>   |                                                                              |==========================================================            |  82%Compiling rjags model...
#> Calling the simulation using the rjags method...
#> Adapting the model for 100 iterations...
#> Burning in the model for 250 iterations...
#> Running the model for 250 iterations...
#> Simulation complete
#> Calculating summary statistics...
#> Finished running the simulation
#>   |                                                                              |============================================================          |  85%Compiling rjags model...
#> Calling the simulation using the rjags method...
#> Adapting the model for 100 iterations...
#> Burning in the model for 250 iterations...
#> Running the model for 250 iterations...
#> Simulation complete
#> Calculating summary statistics...
#> Finished running the simulation
#>   |                                                                              |=============================================================         |  88%Compiling rjags model...
#> Calling the simulation using the rjags method...
#> Adapting the model for 100 iterations...
#> Burning in the model for 250 iterations...
#> Running the model for 250 iterations...
#> Simulation complete
#> Calculating summary statistics...
#> Finished running the simulation
#>   |                                                                              |===============================================================       |  90%Compiling rjags model...
#> Calling the simulation using the rjags method...
#> Adapting the model for 100 iterations...
#> Burning in the model for 250 iterations...
#> Running the model for 250 iterations...
#> Simulation complete
#> Calculating summary statistics...
#> Finished running the simulation
#>   |                                                                              |=================================================================     |  92%Compiling rjags model...
#> Calling the simulation using the rjags method...
#> Adapting the model for 100 iterations...
#> Burning in the model for 250 iterations...
#> Running the model for 250 iterations...
#> Simulation complete
#> Calculating summary statistics...
#> Finished running the simulation
#>   |                                                                              |==================================================================    |  95%Compiling rjags model...
#> Calling the simulation using the rjags method...
#> Adapting the model for 100 iterations...
#> Burning in the model for 250 iterations...
#> Running the model for 250 iterations...
#> Simulation complete
#> Calculating summary statistics...
#> Finished running the simulation
#>   |                                                                              |====================================================================  |  98%Compiling rjags model...
#> Calling the simulation using the rjags method...
#> Adapting the model for 100 iterations...
#> Burning in the model for 250 iterations...
#> Running the model for 250 iterations...
#> Simulation complete
#> Calculating summary statistics...
#> Finished running the simulation
#>   |                                                                              |======================================================================| 100%
#> Calculating summary statistics...
#> Calculating the Gelman-Rubin statistic for 3 variables....
#> Calculating summary statistics...
#> Calculating the Gelman-Rubin statistic for 3 variables....
#> Calculating summary statistics...
#> Calculating the Gelman-Rubin statistic for 3 variables....
#> Calculating summary statistics...
#> Calculating the Gelman-Rubin statistic for 3 variables....
#> Calculating summary statistics...
#> Calculating the Gelman-Rubin statistic for 3 variables....
#> Calculating summary statistics...
#> Calculating the Gelman-Rubin statistic for 3 variables....
#> Calculating summary statistics...
#> Calculating the Gelman-Rubin statistic for 3 variables....
#> Calculating summary statistics...
#> Calculating the Gelman-Rubin statistic for 3 variables....
#> Calculating summary statistics...
#> Calculating the Gelman-Rubin statistic for 3 variables....
#> Calculating summary statistics...
#> Calculating the Gelman-Rubin statistic for 3 variables....

# print model predictions
print(results$predictions)

#> # A tibble: 30 x 2
#>       f1    f2
#>    <dbl> <dbl>
#>  1 0.476 0.696
#>  2 0.416 0.719
#>  3 0.516 0.561
#>  4 0.413 0.689
#>  5 0.541 0.429
#>  6 0.490 0.332
#>  7 0.493 0.629
#>  8 0.505 0.497
#>  9 0.532 0.453
#> 10 0.540 0.400
#> # … with 20 more rows

# print model performance
print(results$performance, width = Inf)

#> # A tibble: 2 x 14
#>   feature max_mpsrf train_tss_mean train_tss_std train_sensitivity_mean
#>   <chr>       <dbl>          <dbl>         <dbl>                  <dbl>
#> 1 f1           1.03          0.527         0.304                  0.716
#> 2 f2           1.04          0.451         0.116                  0.573
#>   train_sensitivity_std train_specificity_mean train_specificity_std
#>                   <dbl>                  <dbl>                 <dbl>
#> 1                 0.137                  0.810                0.183 
#> 2                 0.117                  0.878                0.0629
#>   test_tss_mean test_tss_std test_sensitivity_mean test_sensitivity_std
#>           <dbl>        <dbl>                 <dbl>                <dbl>
#> 1         0.324        0.414                 0.681                0.211
#> 2         0.571        0.274                 0.666                0.288
#>   test_specificity_mean test_specificity_std
#>                   <dbl>                <dbl>
#> 1                 0.643                0.241
#> 2                 0.905                0.155

# }