Fit models to estimate probability of occupancy

Estimate probability of occupancy for a set of features in a set of planning units.

fit_occupancy_models(site_data, site_occupancy_columns,
  site_env_vars_columns, parameters, n_folds = rep(ifelse(nrow(site_data)
  > 1000, 10, 5), length(site_occupancy_columns)),
  site_weight_columns = NULL, n_threads = 1, verbose = FALSE)

Arguments

site_data
sf object with site data.

site_occupancy_columns
character names of numeric columns in the argument to site_data that contain presence/absence data. Each column should correspond to a different feature, and contain binary presence/absence data (zeros or ones) indicating if the feature was detected in a previous survey or not. If a site has not been surveyed before, then missing (NA) values should be used.

site_env_vars_columns
character names of columns in the argument to site_data that contain environmental information for fitting updated occupancy models based on possible survey outcomes. Each column should correspond to a different environmental variable, and contain numeric, factor, or character data. No missing (NA) values are permitted in these columns.

parameters
list of list objects containing the parameters for fitting models for each feature. Ideally, these parameters would be determined using the tune_occupancy_models function. Note that arguments must have a nrounds element (see example below). Valid parameters include: "max_depth", "eta", "nrounds", "lambda", "subsample", "colsample_bytree", and "objective". See documentation for the params argument in xgb.train for more information.

n_folds
numeric number of folds to split the training data into when fitting models for each feature. Defaults to 10 if there are more than 1000 rows in the argument site_data, otherwise defaults to 5. Note that if the maximum number of minority cases (e.g. 1 presence and 1000 absences) is fewer than the number of folds then one minority case is sampled per fold.

site_weight_columns
character name of columns in site_data containing weights for model fitting. These columns must contain numeric values. No missing (NA) values are permitted. Defaults to NULL such that all data are given equal weight when fitting models.

n_threads
integer number of threads to use for fitting models. Defaults to 1.

verbose
logical indicating if information should be printed during computations. Defaults to FALSE.

Value

list object containing the following elements:

predictions

tibble with estimated probabilities of each feature occupying each site. Rows correspond to sites and columns correspond to features.

performance

tibble with model information. Each row corresponds to a different feature and column contain different information. Specifically, the columns contain: (sensitivity) model sensitivities, (specificity) model specificities, and (auc) model area under the curve (AUC) statistics.

Details

The models are fitted using the specified tuning parameters and automatically calibrated scale_pos_weight xgboost parameters (i.e. number of absence divided by number of presences per feature) to account for unbalanced datasets. After the models are fitted, they are used to make probability of occupancy predictions (via xgboost). The average sensitivity, specificity, and area under the curve (AUC) statistics are calculated using the data held are then reported to assess model performance. Note that each feature is run with the same seed to ensure reproducibility.

Examples

# set seed
set.seed(100)

# simulate data for 200 sites, 2 features, and 3 environmental variables
 x <- simulate_site_data(200, 2, 0.5, n_env_vars = 3)

#> ...

# set parameters manually
xgb_params <- list(list(nrounds = 10, objective = "binary:logistic"),
                   list(nrounds = 5, objective = "binary:logistic"))

# fit models
p <- fit_occupancy_models(x, c("f1", "f2"), c("e1", "e2", "e3"),
                          xgb_params)

# preview results
str(p)

#> List of 2
#>  $ predictions:Classes 'tbl_df', 'tbl' and 'data.frame':	200 obs. of  2 variables:
#>   ..$ f1: num [1:200] 0.0405 0.0626 0.4685 0.0522 0.2935 ...
#>   ..$ f2: num [1:200] 0.878 0.878 0.482 0.878 0.482 ...
#>  $ performance:Classes 'tbl_df', 'tbl' and 'data.frame':	2 obs. of  12 variables:
#>   ..$ train_auc_mean        : num [1:2] 1 0.998
#>   ..$ train_auc_std         : num [1:2] 0.000663 0.000864
#>   ..$ train_sensitivity_mean: num [1:2] 0.988 0.975
#>   ..$ train_sensitivity_std : num [1:2] 0.0172 0.0354
#>   ..$ train_specificity_mean: num [1:2] 0.987 0.978
#>   ..$ train_specificity_std : num [1:2] 0.0181 0.0307
#>   ..$ test_auc_mean         : num [1:2] 0.953 0.995
#>   ..$ test_auc_std          : num [1:2] 0.06629 0.00737
#>   ..$ test_sensitivity_mean : num [1:2] 0.808 0.917
#>   ..$ test_sensitivity_std  : num [1:2] 0.0825 0.1179
#>   ..$ test_specificity_mean : num [1:2] 0.958 0.875
#>   ..$ test_specificity_std  : num [1:2] 0.0589 0.1768