Simulated datasets

Simulated data for prioritizing sites for ecological surveys.

Usage

data(sim_features)

data(sim_sites)

Format

sim_sites

sf::sf() object.

sim_features

tibble::tibble() object.

.

Details

The simulated datasets provide data for six sites and three features. The sites can potentially acquired for protected area establishment. However, existing information on the spatial distribution of the features is incomplete. Only some of the sites have existing ecological survey data. To help inform management decisions, species distribution models have been fitted to predict the probability of each species occupying each site.

sim_sites

This object describes the sites and contains the following data: cost of surveying the sites (survey_cost column), cost of acquiring sites for conservation (management_cost column), results from previous ecological surveys (f1, f2, f3 columns), previous survey effort (n1, n2, n3 columns), environmental conditions of the sites (e1, e2 columns), and modeled probability of the features occupying the sites (p1, p2, p3 columns).

sim_features

This object describes the features and contains the following data: the name of each feature (name column), whether each feature should be considered in future surveys (survey column), the sensitivity and specificity of the survey methodology for each the sensitivity and specificity of the species distribution model for each feature (model_sensitivity, model_specificity columns), and the representation target thresholds for each feature (target column).

See also

Examples

# load data
data(sim_sites, sim_features)

# print feature data
print(sim_features, width = Inf)
#> # A tibble: 3 × 7
#>   name  survey survey_sensitivity survey_specificity model_sensitivity
#>   <chr> <lgl>               <dbl>              <dbl>             <dbl>
#> 1 f1    TRUE                0.951              0.854             0.711
#> 2 f2    TRUE                0.990              0.832             0.722
#> 3 f3    TRUE                0.986              0.808             0.772
#>   model_specificity target
#>               <dbl>  <dbl>
#> 1             0.841      2
#> 2             0.806      2
#> 3             0.871      2
# print site data
print(sim_sites, width = Inf)
#> Simple feature collection with 6 features and 13 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 0.02541313 ymin: 0.07851093 xmax: 0.9888107 ymax: 0.717068
#> CRS:           NA
#> # A tibble: 6 × 14
#>   survey_cost management_cost    f1    f2    f3    n1    n2    n3     e1     e2
#>         <dbl>           <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl>  <dbl>
#> 1          19              99     0     0  0        0     0     0  1.13   0.535
#> 2          22              87     0     1  0.25     4     4     4 -1.37  -1.45 
#> 3          13              94     1     1  0        1     1     1  0.155 -0.867
#> 4          19              61     0     0  0        0     0     0 -0.792  1.32 
#> 5           9             105     0     0  0        0     0     0 -0.194  0.238
#> 6          12             136     0     0  0        0     0     0  1.07   0.220
#>      p1    p2    p3               geometry
#>   <dbl> <dbl> <dbl>                <POINT>
#> 1 0.999 0.988 0.21   (0.03529733 0.544939)
#> 2 0.001 0.995 0.152    (0.33276 0.3174416)
#> 3 0.966 1     0.017 (0.6141922 0.07851093)
#> 4 0     0     1     (0.02541313 0.1147132)
#> 5 0.11  0.006 0.831  (0.9888107 0.2152785)
#> 6 1     1     0.082   (0.9038749 0.717068)