Simulate feature data

Simulate feature data for developing simulated survey schemes.

simulate_feature_data(n_features, proportion_of_survey_features = 1)

Arguments

n_features
integer number of features.

proportion_of_survey_features
numeric proportion of features that will be examined in the new surveys. Values must be between zero and one. Defaults to 1 such that all features should be surveyed.

Value

tibble::tibble() object. It contains the following data:

name

character name of each feature.

survey

logical (TRUE / FALSE) values indicating if each feature should be examined in surveys or not.

survey_sensitivity

numeric sensitivity (true positive rate) of the survey methodology for each features.

survey_specificity

numeric specificity (true negative rate) of the survey methodology for each features.

model_sensitivity

numeric specificity (true positive rate) of the occupancy models for each features.

model_specificity

numeric specificity (true negative rate) of the occupancy models for each features.

target

numeric target values used to parametrize the conservation benefit of managing of each feature (defaults to 1).

See also

Examples

# set seed for reproducibility
set.seed(123)

# simulate data
d <- simulate_feature_data(n_features = 5,
                           proportion_of_survey_features = 0.5)
# print data
print(d, width = Inf)
#> # A tibble: 5 × 7
#>   name  survey survey_sensitivity survey_specificity model_sensitivity
#>   <chr> <lgl>               <dbl>              <dbl>             <dbl>
#> 1 f1    FALSE               0.952              0.896             0.790
#> 2 f2    TRUE                0.971              0.845             0.725
#> 3 f3    TRUE                0.986              0.868             0.704
#> 4 f4    FALSE               0.972              0.857             0.733
#> 5 f5    TRUE                0.968              0.810             0.795
#>   model_specificity target
#>               <dbl>  <dbl>
#> 1             0.889      1
#> 2             0.869      1
#> 3             0.864      1
#> 4             0.899      1
#> 5             0.866      1