surveyvoi: Survey Value of Information

Decision support tool for prioritizing sites for ecological surveys based on their potential to improve plans for conserving biodiversity (e.g. plans for establishing protected areas). Given a set of sites that could potentially be acquired for conservation management – wherein some sites have previously been surveyed and other sites have not – it can be used to generate and evaluate plans for additional surveys. Specifically, plans for ecological surveys can be generated using various conventional approaches (e.g. maximizing expected species richness, geographic coverage, diversity of sampled environmental conditions) and by maximizing value of information. After generating plans for surveys, they can also be evaluated using value of information analysis.

Details

Please note that several functions depend on the 'Gurobi' optimization software (available from https://www.gurobi.com) and the gurobi R package (installation instructions available online for Linux, Windows, and Mac OS). Additionally, the JAGS software (available from https://mcmc-jags.sourceforge.io/) is required to fit hierarchical generalized linear models.

See also

Author

Package authors:

Jeffrey O. Hanson jeffrey.hanson@uqconnect.edu.au ORCID

Iadine Chadès iadine.chades@csiro.au ORCID

Emma J. Hudgins emma.hudgins@mail.mcgill.ca ORCID

Joseph R. Bennett joseph.bennett@carleton.ca ORCID