Approximate expected value of the decision given current information

Calculate the expected value of the conservation management decision given current information. This metric describes the value of the management decision that is expected when the decision maker is limited to existing biodiversity data (i.e. survey data and environmental niche models).

approx_evdci(
  site_data,
  feature_data,
  site_occupancy_columns,
  site_probability_columns,
  site_management_cost_column,
  feature_survey_sensitivity_column,
  feature_survey_specificity_column,
  feature_model_sensitivity_column,
  feature_model_specificity_column,
  feature_preweight_column,
  feature_postweight_column,
  feature_target_column,
  total_budget,
  site_management_locked_in_column = NULL,
  prior_matrix = NULL,
  n_approx_obj_fun_points = 1000,
  optimality_gap = 0,
  n_approx_replicates = 100,
  n_approx_states_per_replicate = 1000,
  method_approx_states = "weighted_without_replacement",
  seed = 500
)

Arguments

site_data
sf object with site data.

feature_data
data.frame object with feature data.

site_occupancy_columns
character names of numeric columns in the argument to site_data that contain presence/absence data. Each column should correspond to a different feature, and contain binary presence/absence data (zeros or ones) indicating if the feature was detected in a previous survey or not. If a site has not been surveyed before, then missing (NA) values should be used.

site_probability_columns
character names of numeric columns in the argument to site_data that contain modelled probabilities of occupancy for each feature in each site. Each column should correspond to a different feature, and contain probability data (values between zero and one). No missing (NA) values are permitted in these columns.

site_management_cost_column
character name of column in the argument to site_data that contains costs for managing each site for conservation. This column should have numeric values that are equal to or greater than zero. No missing (NA) values are permitted in this column.

feature_survey_sensitivity_column
character name of the column in the argument to feature_data that contains probability of future surveys correctly detecting a presence of each feature in a given site (i.e. the sensitivity of the survey methodology). This column should have numeric values that are between zero and one. No missing (NA) values are permitted in this column.

feature_survey_specificity_column
character name of the column in the argument to feature_data that contains probability of future surveys correctly detecting an absence of each feature in a given site (i.e. the specificity of the survey methodology). This column should have numeric values that are between zero and one. No missing (NA) values are permitted in this column.

feature_model_sensitivity_column
character name of the column in the argument to feature_data that contains probability of the initial models correctly predicting a presence of each feature in a given site (i.e. the sensitivity of the models). This column should have numeric values that are between zero and one. No missing (NA) values are permitted in this column. This should ideally be calculated using fit_occupancy_models.

feature_model_specificity_column
character name of the column in the argument to feature_data that contains probability of the initial models correctly predicting an absence of each feature in a given site (i.e. the specificity of the models). This column should have numeric values that are between zero and one. No missing (NA) values are permitted in this column. This should ideally be calculated using fit_occupancy_models.

feature_preweight_column
character name of the column in the argument to feature_data that contains the \(preweight\) values used to parametrize the conservation benefit of managing of each feature. This column should have numeric values that are equal to or greater than zero. No missing (NA) values are permitted in this column.

feature_postweight_column
character name of the column in the argument to feature_data that contains the \(postweight\) values used to parametrize the conservation benefit of managing of each feature. This column should have numeric values that are equal to or greater than zero. No missing (NA) values are permitted in this column.

feature_target_column
character name of the column in the argument to feature_data that contains the \(target\) values used to parametrize the conservation benefit of managing of each feature. This column should have numeric values that are equal to or greater than zero. No missing (NA) values are permitted in this column.

total_budget
numeric maximum expenditure permitted for conducting surveys and managing sites for conservation.

site_management_locked_in_column
character name of the column in the argument to site_data that contains logical (TRUE / FALSE) values indicating which sites should be locked in for (TRUE) being managed for conservation or (FALSE) not. No missing (NA) values are permitted in this column. This is useful if some sites have already been earmarked for conservation, or if some sites are already being managed for conservation. Defaults to NULL such that no sites are locked in.

prior_matrix
numeric matrix containing the prior probability of each feature occupying each site. Rows correspond to features, and columns correspond to sites. Defaults to NULL such that prior data is calculated automatically using prior_probability_matrix.

n_approx_obj_fun_points
integer number of points to use for approximating the piecewise-linear components of the objective function. Greater values result in more precise calculations, but also incur more computational costs. Defaults to 1000.

optimality_gap
numeric relative optimality gap for generating conservation prioritizations. A value of zero indicates that prioritizations must be solved to optimality. A value of 0.1 indicates prioritizations must be within 10% of optimality. Defaults to 0.

n_approx_replicates
integer number of replicates to use for approximating the expected value calculations. Defaults to 100.

n_approx_states_per_replicate
integer number of states to use per replicate for approximating the expected value of a given management action. This number must be smaller than or equal to the total number of presence absence states in the system (i.e. n_states(nrow(site_data), nrow(feature_data))) Defaults to 1000.

method_approx_states
character name of method that is used to sample states for approximating the expected value calculations. Available options are: "uniform_with_replacement", "uniform_without_replacement", "weighted_with_replacement", "weighted_without_replacement". Uniform sampling methods have an equal chance of returning each state, and weighted sampling methods are more likely to return states with a higher prior probability of occurring. Defaults to "weighted_without_replacement".

seed
integer random seed used for generating states. Defaults to 500.

Value

numeric vector containing the estimate for each replicate.

See also

Examples

# set seeds for reproducibility
library(RandomFields)

#> Loading required package: sp

#> Loading required package: RandomFieldsUtils

#> 
#> Attaching package: ‘RandomFields’

#> The following object is masked from ‘package:RandomFieldsUtils’:
#> 
#>     RFoptions

set.seed(123)
RFoptions(seed = 123)

# simulate data
site_data <- simulate_site_data(n_sites = 5, n_features = 2, prop = 0.5)

#> ...

feature_data <- simulate_feature_data(n_features = 2, prop = 1)

# preview simulated data
print(site_data)

#> Simple feature collection with 5 features and 9 fields
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: 0.2875775 ymin: 0.0455565 xmax: 0.9404673 ymax: 0.892419
#> epsg (SRID):    NA
#> proj4string:    NA
#> # A tibble: 5 x 10
#>   survey_cost management_cost     e1     e2     e3    f1    f2    p1     p2
#>         <dbl>           <dbl>  <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl>  <dbl>
#> 1        8.13            73.9 -0.365  0.642  1.08      1     1 0.950 0.940 
#> 2        6.07            69.7 -0.303 -0.382 -0.838     0     0 0.229 0.362 
#> 3        3.74            62.0  1.77   1.42   1.11     NA    NA 0.789 0.0228
#> 4        6.82            71.8 -0.404 -0.771 -0.732    NA    NA 0.157 0.552 
#> 5        5.22            70.7 -0.697 -0.910 -0.614     1     1 0.201 0.798 
#> # … with 1 more variable: geometry <POINT>

print(feature_data)

#> # A tibble: 2 x 9
#>   name  survey survey_sensitiv… survey_specific… model_sensitivi…
#>   <chr> <lgl>             <dbl>            <dbl>            <dbl>
#> 1 f1    TRUE              0.955            0.822            0.775
#> 2 f2    TRUE              0.961            0.829            0.723
#> # … with 4 more variables: model_specificity <dbl>, preweight <dbl>,
#> #   postweight <dbl>, target <dbl>

# set total budget for managing sites for conservation
total_budget <- sum(site_data$management_cost) * 0.5

# calculate expected value of management decision given current information
# using approximate method with 100 replicates and 50 states per replicate
ev_prime_current <- approx_evdci(
    site_data, feature_data, c("f1", "f2"), c("p1", "p2"),
    "management_cost", "survey_sensitivity", "survey_specificity",
    "model_sensitivity", "model_specificity",
    "preweight", "postweight", "target",
    total_budget, n_approx_replicates = 100,
    n_approx_states_per_replicate = 50)

# print approximate value
print(ev_prime_current)

#>   [1] 27.16712 27.23450 27.31186 27.31336 27.16887 27.36405 27.20167 27.36456
#>   [9] 27.24082 27.16902 27.22147 27.21514 27.31831 27.23862 27.28213 27.19937
#>  [17] 27.18638 27.24782 27.26162 27.16070 27.17609 27.16298 27.21706 27.18931
#>  [25] 27.24178 27.19772 27.25706 27.18566 27.14838 27.29333 27.28450 27.18872
#>  [33] 27.22379 27.09731 27.25328 27.14261 27.27162 27.23463 27.22539 27.15424
#>  [41] 27.19488 27.10314 27.22263 27.16217 27.15964 27.27873 27.16303 27.18776
#>  [49] 27.23535 27.20835 27.13136 27.21959 27.36306 27.14237 27.12986 27.19522
#>  [57] 27.21227 27.10798 27.18715 27.27769 27.12803 27.28120 27.25832 27.17789
#>  [65] 27.09634 27.21861 27.19615 27.19496 27.19873 27.15497 27.18141 27.25063
#>  [73] 27.24024 27.12954 27.30984 27.18960 27.28053 27.24902 27.34193 27.32329
#>  [81] 27.28894 27.24825 27.19682 27.18428 27.19049 27.20612 27.15817 27.16242
#>  [89] 27.30135 27.12682 27.17040 27.16666 27.23513 27.16413 27.12130 27.09635
#>  [97] 27.26411 27.17272 27.38791 27.20376