Approximately optimal survey scheme

Find the optimal survey scheme that maximizes return on investment using value of information analyses. This function uses the approximation method for calculating the expected value of the decision given a survey scheme.

approx_optimal_survey_scheme(
  site_data,
  feature_data,
  site_occupancy_columns,
  site_probability_columns,
  site_env_vars_columns,
  site_management_cost_column,
  site_survey_cost_column,
  feature_survey_column,
  feature_survey_sensitivity_column,
  feature_survey_specificity_column,
  feature_model_sensitivity_column,
  feature_model_specificity_column,
  feature_preweight_column,
  feature_postweight_column,
  feature_target_column,
  total_budget,
  survey_budget,
  xgb_parameters,
  site_management_locked_in_column = NULL,
  site_survey_locked_out_column = NULL,
  prior_matrix = NULL,
  optimality_gap = 0,
  site_weight_columns = NULL,
  xgb_n_folds = rep(5, nrow(feature_data)),
  n_approx_replicates = 100,
  n_approx_outcomes_per_replicate = 10000,
  method_approx_outcomes = "weighted_without_replacement",
  seed = 500,
  n_threads = 1
)

Arguments

site_data
sf object with site data.

feature_data
data.frame object with feature data.

site_occupancy_columns
character names of numeric columns in the argument to site_data that contain presence/absence data. Each column should correspond to a different feature, and contain binary presence/absence data (zeros or ones) indicating if the feature was detected in a previous survey or not. If a site has not been surveyed before, then missing (NA) values should be used.

site_probability_columns
character names of numeric columns in the argument to site_data that contain modelled probabilities of occupancy for each feature in each site. Each column should correspond to a different feature, and contain probability data (values between zero and one). No missing (NA) values are permitted in these columns.

site_env_vars_columns
character names of columns in the argument to site_data that contain environmental information for fitting updated occupancy models based on possible survey outcomes. Each column should correspond to a different environmental variable, and contain numeric, factor, or character data. No missing (NA) values are permitted in these columns.

site_management_cost_column
character name of column in the argument to site_data that contains costs for managing each site for conservation. This column should have numeric values that are equal to or greater than zero. No missing (NA) values are permitted in this column.

site_survey_cost_column
character name of column in the argument to site_data that contains costs for surveying each site. This column should have numeric values that are equal to or greater than zero. No missing (NA) values are permitted in this column.

feature_survey_column
character name of the column in the argument to feature_data that contains logical (TRUE / FALSE) values indicating if the feature will be surveyed in the planned surveys or not. Note that considering additional features will rapidly increase computational burden, and so it is only recommended to consider features that are of specific conservation interest. No missing (NA) values are permitted in this column.

feature_survey_sensitivity_column
character name of the column in the argument to feature_data that contains probability of future surveys correctly detecting a presence of each feature in a given site (i.e. the sensitivity of the survey methodology). This column should have numeric values that are between zero and one. No missing (NA) values are permitted in this column.

feature_survey_specificity_column
character name of the column in the argument to feature_data that contains probability of future surveys correctly detecting an absence of each feature in a given site (i.e. the specificity of the survey methodology). This column should have numeric values that are between zero and one. No missing (NA) values are permitted in this column.

feature_model_sensitivity_column
character name of the column in the argument to feature_data that contains probability of the initial models correctly predicting a presence of each feature in a given site (i.e. the sensitivity of the models). This column should have numeric values that are between zero and one. No missing (NA) values are permitted in this column. This should ideally be calculated using fit_occupancy_models.

feature_model_specificity_column
character name of the column in the argument to feature_data that contains probability of the initial models correctly predicting an absence of each feature in a given site (i.e. the specificity of the models). This column should have numeric values that are between zero and one. No missing (NA) values are permitted in this column. This should ideally be calculated using fit_occupancy_models.

feature_preweight_column
character name of the column in the argument to feature_data that contains the \(preweight\) values used to parametrize the conservation benefit of managing of each feature. This column should have numeric values that are equal to or greater than zero. No missing (NA) values are permitted in this column.

feature_postweight_column
character name of the column in the argument to feature_data that contains the \(postweight\) values used to parametrize the conservation benefit of managing of each feature. This column should have numeric values that are equal to or greater than zero. No missing (NA) values are permitted in this column.

feature_target_column
character name of the column in the argument to feature_data that contains the \(target\) values used to parametrize the conservation benefit of managing of each feature. This column should have numeric values that are equal to or greater than zero. No missing (NA) values are permitted in this column.

total_budget
numeric maximum expenditure permitted for conducting surveys and managing sites for conservation.

survey_budget
numeric maximum expenditure permitted for conducting surveys.

xgb_parameters
list of list objects containing the parameters for fitting models for each feature. See documentation for the params argument in xgb.train for available parameters. Ideally, these parameters would be determined using the fit_occupancy_models function. Note that arguments must have "nrounds", "objective", "scale_pos_weight" elements (see example below).

site_management_locked_in_column
character name of the column in the argument to site_data that contains logical (TRUE / FALSE) values indicating which sites should be locked in for (TRUE) being managed for conservation or (FALSE) not. No missing (NA) values are permitted in this column. This is useful if some sites have already been earmarked for conservation, or if some sites are already being managed for conservation. Defaults to NULL such that no sites are locked in.

site_survey_locked_out_column
character name of the column in the argument to site_data that contains logical (TRUE / FALSE) values indicating which sites should be locked out (TRUE) from being selected for future surveys or (FALSE) not. No missing (NA) values are permitted in this column. This is useful if some sites will never be considered for future surveys (e.g. because they are too costly to survey, or have a low chance of containing the target species). Defaults to NULL such that no sites are locked out.

prior_matrix
numeric matrix containing the prior probability of each feature occupying each site. Rows correspond to features, and columns correspond to sites. Defaults to NULL such that prior data is calculated automatically using prior_probability_matrix.

optimality_gap
numeric relative optimality gap for generating conservation prioritizations. A value of zero indicates that prioritizations must be solved to optimality. A value of 0.1 indicates prioritizations must be within 10% of optimality. Defaults to 0.

site_weight_columns
character name of columns in site_data containing weights for model fitting for each feature. These columns must contain numeric values greater than or equal to zero. No missing (NA) values are permitted. Defaults to NULL such that all data are given equal weight when fitting models.

xgb_n_folds
integer vector containing the number of k-fold cross-validation folds to use for fitting models and assessing model performance for each feature. Ideally, the number of folds should be exactly the same as the number used for tuning the model parameters (i.e. same parameter to the n_folds argument in fit_occupancy_models when generating parameters for xgb_parameters).

n_approx_replicates
integer number of replicates to use for approximating the expected value calculations. Defaults to 100.

n_approx_outcomes_per_replicate
integer number of outcomes to #' use per replicate for approximation calculations. Defaults to 10000.

method_approx_outcomes
character name of method that is used to sample outcomes for approximating the expected value calculations. Available options are: "uniform_with_replacement", "uniform_without_replacement", "weighted_with_replacement", "weighted_without_replacement". Uniform sampling methods have an equal chance of returning each state, and weighted sampling methods are more likely to return outcomes with a higher prior probability of occurring. Defaults to "weighted_without_replacement".

seed
integer state of the random number generator for partitioning data into folds cross-validation and fitting xgboost models. It is also used for generating outcomes. This parameter must remain the same to compare results from different functions using the approximation methods. Defaults to 500.

n_threads
integer number of threads to use for computation.

Value

matrix of logical (TRUE/ FALSE) values indicating if a site is selected in the scheme or not. Columns correspond to sites, and rows correspond to different schemes. If there is only one optimal survey scheme then the matrix will only contain a single row.

This matrix also has a numeric "ev" attribute that contains a matrix with the approximate expected values. Within this attribute, each row corresponds to a different survey scheme and each column corresponds to a different replicate.

Details

The "approximately" optimal survey scheme is determined using a brute-force algorithm. Initially, all feasible (valid) survey schemes are identified given the survey costs and the survey budget (using feasible_survey_schemes. Next, the expected value of each and every feasible survey scheme is approximated (using approx_evdsi). Finally, the greatest expected value is identified, and all survey schemes that share this greatest expected value are returned. Due to the nature of this algorithm, it can take a very long time to complete.

Examples

# set seeds for reproducibility
library(RandomFields)
set.seed(123)
RFoptions(seed = 201)

# simulate data
site_data <- simulate_site_data(n_sites = 11, n_features = 2, prop = 0.8)

#> ...

feature_data <- simulate_feature_data(n_features = 2, prop = 1)

# preview simulated data
print(site_data)

#> Simple feature collection with 11 features and 9 fields
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: 0.0455565 ymin: 0.04205953 xmax: 0.9568333 ymax: 0.9545036
#> CRS:            NA
#> # A tibble: 11 x 10
#>    survey_cost management_cost      e1      e2       e3    f1    f2     p1
#>          <dbl>           <dbl>   <dbl>   <dbl>    <dbl> <dbl> <dbl>  <dbl>
#>  1       12.2             64.6  1.01   -0.958   0.898       0     0 0.0914
#>  2       11.8             66.6  0.320   0.0747  0.262      NA    NA 0.491 
#>  3       11.1             64.4  0.924  -0.544   1.10       NA    NA 0.239 
#>  4       10.4             78.1 -2.10   -0.296  -1.60       NA    NA 0.676 
#>  5        7.27            65.4  0.177   1.24   -0.326      NA    NA 0.848 
#>  6       11.4             72.5  1.20   -1.35    0.00343    NA    NA 0.0132
#>  7       10.5             73.8 -0.924  -0.643  -1.63       NA    NA 0.157 
#>  8        8.93            75.0 -1.08   -0.654  -0.659      NA    NA 0.407 
#>  9        8.14            71.4  0.0140  1.73    0.764       1     1 0.984 
#> 10        9.09            72.9  0.310   1.20    1.25       NA    NA 0.961 
#> 11       10.8             66.4  0.141   0.191  -0.0521     NA    NA 0.524 
#> # ... with 2 more variables: p2 <dbl>, geometry <POINT>

print(feature_data)

#> # A tibble: 2 x 9
#>   name  survey survey_sensitiv~ survey_specific~ model_sensitivi~
#>   <chr> <lgl>             <dbl>            <dbl>            <dbl>
#> 1 f1    TRUE              0.956            0.892            0.787
#> 2 f2    TRUE              0.963            0.863            0.702
#> # ... with 4 more variables: model_specificity <dbl>, preweight <dbl>,
#> #   postweight <dbl>, target <dbl>

# set total budget for managing sites for conservation
# (i.e. 50% of the cost of managing all sites)
total_budget <- sum(site_data$management_cost) * 0.5

# set total budget for surveying sites for conservation
# (i.e. 10% of the cost of managing all sites)
survey_budget <- sum(site_data$survey_cost) * 0.1

# define xgboost tuning parameters
xgb_parameters <-
 list(list(objective = "binary:logistic", scale_pos_weight = 1,
           nrounds = 8, eta = 0.1))[rep(1, 2)]

# find optimal survey scheme using approximate method
# (using 10 replicates so that this example completes relatively quickly)
approx_opt_survey <- approx_optimal_survey_scheme(
  site_data, feature_data,
  c("f1", "f2"), c("p1", "p2"), c("e1", "e2", "e3"),
  "management_cost", "survey_cost",
  "survey", "survey_sensitivity", "survey_specificity",
  "model_sensitivity", "model_specificity",
  "preweight", "postweight", "target",
  total_budget, survey_budget, xgb_parameters,
  n_approx_replicates = 10)

#> Error in as(rcpp_feasible_actions_ilp_matrix(costs), "dgCMatrix"): no method or default for coercing "matrix" to "dgCMatrix"

# print result
print(approx_opt_survey)

#> Error in print(approx_opt_survey): object 'approx_opt_survey' not found