Fit models to estimate probability of occupancy

Estimate probability of occupancy for a set of features in a set of planning units.

fit_occupancy_models(
  site_data,
  site_occupancy_columns,
  site_env_vars_columns,
  parameters,
  early_stopping_rounds = 100,
  n_rounds = 1000,
  n_folds = rep(5, length(site_occupancy_columns)),
  n_random_search_iterations = 10000,
  site_weight_columns = NULL,
  n_threads = 1,
  seed = 500,
  verbose = FALSE
)

Arguments

site_data
sf object with site data.

site_occupancy_columns
character names of numeric columns in the argument to site_data that contain presence/absence data. Each column should correspond to a different feature, and contain binary presence/absence data (zeros or ones) indicating if the feature was detected in a previous survey or not. If a site has not been surveyed before, then missing (NA) values should be used.

site_env_vars_columns
character names of columns in the argument to site_data that contain environmental information for fitting updated occupancy models based on possible survey outcomes. Each column should correspond to a different environmental variable, and contain numeric, factor, or character data. No missing (NA) values are permitted in these columns.

parameters
list object containing the candidate parameter values for fitting models. Valid parameters include: "max_depth", "eta", "lambda", "min_child_weight", "subsample", "colsample_by_tree", "objective". See documentation for the params argument in xgb.train for more information.

early_stopping_rounds
numeric model rounds for parameter tuning. See xgboost for more information. Defaults to 100.

n_rounds
numeric model rounds for model fitting See xgboost for more information. Defaults to 1000.

n_folds
numeric number of folds to split the training data into when fitting models for each feature. Defaults to 10 if there are more than 1000 rows in the argument site_data, otherwise defaults to 5. Note that if the maximum number of minority cases (e.g. 1 presence and 1000 absences) is fewer than the number of folds then one minority case is sampled per fold.

n_random_search_iterations
numeric number of random search iterations to use when tuning model parameters. Defaults to 10000.

site_weight_columns
character name of columns in site_data containing weights for model fitting. These columns must contain numeric values. No missing (NA) values are permitted. Defaults to NULL such that all data are given equal weight when fitting models.

n_threads
integer number of threads to use for parameter tuning. Defaults to 1.

seed
integer initial random number generator state for model fitting. Defaults to 500.

verbose
logical indicating if information should be printed during computations. Defaults to FALSE.

Value

list object containing:

parameters

list of list objects containing the best tuning parameters for each feature.

predictions

tibble object containing predictions for each feature.

performance

tibble object containing the performance of the best cross-validation models for each feature.

Details

This function (i) prepares the data for model fitting, (ii) calibrates the tuning parameters for model fitting, (iii) generate predictions using the best found tuning parameters, and (iv) assess the performance of the best supported models. These analyses are performed separately for each feature. For a given feature:

The data are prepared for model fitting by partitioning the data using k-fold cross-validation (set via argument to n_folds). The training and evaluation folds are constructed in such a manner as to ensure that each training and evaluation fold contains at least one presence and one absence observation.

A random search method is used to tune the model parameters. The candidate values for each parameter (specified via parameters) are used to generate a full set of parameter combinations, and then a subset of these parameter combinations is randomly selected for the tuning procedure (specified via n_random_search_iterations). To account for unbalanced datasets, the scale_pos_weight xgboost parameter is calculated as the mean value across each of the training folds (i.e. number of absence divided by number of presences per feature). For a given parameter combination, models are fit using k-fold cross- validation (via xgb.cv) -- using the previously mentioned training and evaluation folds -- and the average area under the curve (AUC) statistic calculated using the data held out from each fold is used to quantify the performance. These models are also fitted using the early_stopping_rounds parameter to reduce time-spent tuning models. If relevant, they are also fitted using the supplied weights (per by the argument to site_weights_data).After exploring the subset of parameter combinations, the best parameter combination is identified, and the associated parameter values and models are stored for later use.

The cross-validation models associated with the best parameter combination are used to generate predict the average probability that the feature occupies each site. These predictions include sites that have been surveyed before, and also sites that have not been surveyed before.

The performance of the cross-validation models is evaluated. Specifically, the AUC, sensitivity, and specificity statistics are calculated (if relevant, weighted by the argument to site_weights_data). These performance values are calculated using the models' evaluation folds.

Examples

# set seeds for reproducibility
library(RandomFields)
set.seed(123)
RFoptions(seed = 123)

# simulate data for 200 sites, 2 features, and 3 environmental variables
 x <- simulate_site_data(200, 2, 0.5, n_env_vars = 3)

#> ...

# create list of possible tuning parameters for modelling
all_parameters <- list(max_depth = seq(1, 10, 1),
                       eta = seq(0.1, 0.5, 0.1),
                       lambda = 10 ^ seq(-1.0, 0.0, 0.25),
                       subsample = seq(0.5, 1.0, 0.1),
                       colsample_bytree = seq(0.4, 1.0, 0.1),
                       objective = "binary:logistic")

# fit models
# note that we use 10 random search iterations here so that the example
# finishes quickly, you would probably want something like 1000+
results <- fit_occupancy_models(
   x, paste0("f", seq_len(2)), paste0("e", seq_len(3)),
   n_folds = rep(5, 2), n_random_search_iterations = 10,
   early_stopping_rounds = 100, parameters = all_parameters, n_threads = 1)

# print best found model parameters
print(results$parameters)

#> [[1]]
#> [[1]]$max_depth
#> [1] 2
#> 
#> [[1]]$eta
#> [1] 0.2
#> 
#> [[1]]$nrounds
#> [1] 25
#> 
#> [[1]]$scale_pos_weight
#> [1] 1.089492
#> 
#> [[1]]$lambda
#> [1] 0.3162278
#> 
#> [[1]]$subsample
#> [1] 0.6
#> 
#> [[1]]$colsample_bytree
#> [1] 1
#> 
#> [[1]]$objective
#> [1] "binary:logistic"
#> 
#> 
#> [[2]]
#> [[2]]$max_depth
#> [1] 2
#> 
#> [[2]]$eta
#> [1] 0.1
#> 
#> [[2]]$nrounds
#> [1] 19
#> 
#> [[2]]$scale_pos_weight
#> [1] 0.9653068
#> 
#> [[2]]$lambda
#> [1] 0.3162278
#> 
#> [[2]]$subsample
#> [1] 0.7
#> 
#> [[2]]$colsample_bytree
#> [1] 0.5
#> 
#> [[2]]$objective
#> [1] "binary:logistic"
#> 
#> 

# print model predictions
print(results$predictions)

#> # A tibble: 200 x 2
#>        f1    f2
#>     <dbl> <dbl>
#>  1 0.946  0.577
#>  2 0.141  0.636
#>  3 0.421  0.697
#>  4 0.0183 0.215
#>  5 0.0355 0.215
#>  6 0.679  0.828
#>  7 0.0813 0.481
#>  8 0.132  0.232
#>  9 0.541  0.304
#> 10 0.732  0.362
#> # ... with 190 more rows

# print model performance
print(results$performance)

#> # A tibble: 2 x 13
#>   feature train_auc_mean train_auc_std train_sensitivi~ train_sensitivi~
#>   <chr>            <dbl>         <dbl>            <dbl>            <dbl>
#> 1 f1               0.977       0.00475            0.932           0.0232
#> 2 f2               0.961       0.0130             0.862           0.0307
#> # ... with 8 more variables: train_specificity_mean <dbl>,
#> #   train_specificity_std <dbl>, test_auc_mean <dbl>, test_auc_std <dbl>,
#> #   test_sensitivity_mean <dbl>, test_sensitivity_std <dbl>,
#> #   test_specificity_mean <dbl>, test_specificity_std <dbl>