Fit models to estimate probability of occupancy

Estimate probability of occupancy for a set of features in a set of planning units.

fit_occupancy_models(
  site_data,
  site_occupancy_columns,
  site_env_vars_columns,
  xgb_tuning_parameters,
  xgb_early_stopping_rounds = rep(100, length(site_occupancy_columns)),
  xgb_n_rounds = rep(1000, length(site_occupancy_columns)),
  xgb_n_folds = rep(5, length(site_occupancy_columns)),
  site_weight_columns = NULL,
  n_threads = 1,
  seed = 500,
  verbose = FALSE
)

Arguments

site_data
sf object with site data.

site_occupancy_columns
character names of numeric columns in the argument to site_data that contain presence/absence data. Each column should correspond to a different feature, and contain binary presence/absence data (zeros or ones) indicating if the feature was detected in a previous survey or not. If a site has not been surveyed before, then missing (NA) values should be used.

site_env_vars_columns
character names of columns in the argument to site_data that contain environmental information for fitting updated occupancy models based on possible survey outcomes. Each column should correspond to a different environmental variable, and contain numeric, factor, or character data. No missing (NA) values are permitted in these columns.

xgb_tuning_parameters
list object containing the candidate parameter values for fitting models. Valid parameters include: "max_depth", "eta", "lambda", "min_child_weight", "subsample", "colsample_by_tree", "objective". See documentation for the params argument in xgb.train for more information.

xgb_early_stopping_rounds
numeric model rounds for parameter tuning. See xgboost for more information. Defaults to 100 for each feature.

xgb_n_rounds
numeric model rounds for model fitting See xgboost for more information. Defaults to 1000 for each feature.

xgb_n_folds
numeric number of folds to split the training data into when fitting models for each feature. Defaults to 5 for each feature.

site_weight_columns
character name of columns in site_data containing weights for model fitting. These columns must contain numeric values. No missing (NA) values are permitted. Defaults to NULL such that all data are given equal weight when fitting models.

n_threads
integer number of threads to use for parameter tuning. Defaults to 1.

seed
integer initial random number generator state for model fitting. Defaults to 500.

verbose
logical indicating if information should be printed during computations. Defaults to FALSE.

Value

list object containing:

parameters

list of list objects containing the best tuning parameters for each feature.

predictions

tibble object containing predictions for each feature.

performance

tibble object containing the performance of the best models for each feature. It contains the following columns:
feature

name of the feature.

train_tss_mean

mean TSS statistic for models calculated using training data in cross-validation.

train_tss_std

standard deviation in TSS statistics for models calculated using training data in cross-validation.

train_sensitivity_mean

mean sensitivity statistic for models calculated using training data in cross-validation.

train_sensitivity_std

standard deviation in sensitivity statistics for models calculated using training data in cross-validation.

train_specificity_mean

mean specificity statistic for models calculated using training data in cross-validation.

train_specificity_std

standard deviation in specificity statistics for models calculated using training data in cross-validation.

test_tss_mean

mean TSS statistic for models calculated using test data in cross-validation.

test_tss_std

standard deviation in TSS statistics for models calculated using test data in cross-validation.

test_sensitivity_mean

mean sensitivity statistic for models calculated using test data in cross-validation.

test_sensitivity_std

standard deviation in sensitivity statistics for models calculated using test data in cross-validation.

test_specificity_mean

mean specificity statistic for models calculated using test data in cross-validation.

test_specificity_std

standard deviation in specificity statistics for models calculated using test data in cross-validation.

Details

This function (i) prepares the data for model fitting, (ii) calibrates the tuning parameters for model fitting, (iii) generate predictions using the best found tuning parameters, and (iv) assess the performance of the best supported models. These analyses are performed separately for each feature. For a given feature:

The data are prepared for model fitting by partitioning the data using k-fold cross-validation (set via argument to xgb_n_folds). The training and evaluation folds are constructed in such a manner as to ensure that each training and evaluation fold contains at least one presence and one absence observation.

A grid search method is used to tune the model parameters. The candidate values for each parameter (specified via parameters) are used to generate a full set of parameter combinations, and these parameter combinations are subsequently used for tuning the models. To account for unbalanced datasets, the scale_pos_weight xgboost parameter is calculated as the mean value across each of the training folds (i.e. number of absence divided by number of presences per feature). For a given parameter combination, models are fit using k-fold cross- validation (via xgb.cv) -- using the previously mentioned training and evaluation folds -- and the True Skill Statistic (TSS) calculated using the data held out from each fold is used to quantify the performance (i.e. "test_tss_mean" column in output). These models are also fitted using the early_stopping_rounds parameter to reduce time-spent tuning models. If relevant, they are also fitted using the supplied weights (per by the argument to site_weights_data). After exploring the full set of parameter combinations, the best parameter combination is identified, and the associated parameter values and models are stored for later use.

The cross-validation models associated with the best parameter combination are used to generate predict the average probability that the feature occupies each site. These predictions include sites that have been surveyed before, and also sites that have not been surveyed before.

The performance of the cross-validation models is evaluated. Specifically, the TSS, sensitivity, and specificity statistics are calculated (if relevant, weighted by the argument to site_weights_data). These performance values are calculated using the models' training and evaluation folds.

Examples

# set seeds for reproducibility
library(RandomFields)
set.seed(123)
RFoptions(seed = 123)

# simulate data for 200 sites, 2 features, and 3 environmental variables
 x <- simulate_site_data(200, 2, 0.5, n_env_vars = 3)

#> ...

# create list of possible tuning parameters for modelling
parameters <- list(eta = seq(0.1, 0.5, length.out = 3),
                   lambda = 10 ^ seq(-1.0, 0.0, length.out = 3),
                   objective = "binary:logistic")

# fit models
# note that we use 10 random search iterations here so that the example
# finishes quickly, you would probably want something like 1000+
results <- fit_occupancy_models(
   x, paste0("f", seq_len(2)), paste0("e", seq_len(3)),
   xgb_n_folds = rep(5, 2), xgb_early_stopping_rounds = rep(100, 2),
   xgb_tuning_parameters = parameters, n_threads = 1)

# print best found model parameters
print(results$parameters)

#> [[1]]
#> [[1]]$eta
#> [1] 0.5
#> 
#> [[1]]$lambda
#> [1] 0.1
#> 
#> [[1]]$objective
#> [1] "binary:logistic"
#> 
#> [[1]]$nrounds
#> [1] 15
#> 
#> [[1]]$scale_pos_weight
#> [1] 1.040897
#> 
#> 
#> [[2]]
#> [[2]]$eta
#> [1] 0.1
#> 
#> [[2]]$lambda
#> [1] 1
#> 
#> [[2]]$objective
#> [1] "binary:logistic"
#> 
#> [[2]]$nrounds
#> [1] 31
#> 
#> [[2]]$scale_pos_weight
#> [1] 1.083536
#> 
#> 

# print model predictions
print(results$predictions)

#> # A tibble: 200 x 2
#>         f1     f2
#>      <dbl>  <dbl>
#>  1 0.987   0.773 
#>  2 0.0202  0.600 
#>  3 0.0506  0.787 
#>  4 0.00666 0.0473
#>  5 0.0442  0.0488
#>  6 0.706   0.899 
#>  7 0.0487  0.228 
#>  8 0.133   0.0724
#>  9 0.799   0.0501
#> 10 0.925   0.165 
#> # ... with 190 more rows

# print model performance
print(results$performance)

#> # A tibble: 2 x 13
#>   feature train_tss_mean train_tss_std train_sensitivi~ train_sensitivi~
#>   <chr>            <dbl>         <dbl>            <dbl>            <dbl>
#> 1 f1               0.960        0.0286            0.980           0.0114
#> 2 f2               0.931        0.0278            0.974           0.0256
#> # ... with 8 more variables: train_specificity_mean <dbl>,
#> #   train_specificity_std <dbl>, test_tss_mean <dbl>, test_tss_std <dbl>,
#> #   test_sensitivity_mean <dbl>, test_sensitivity_std <dbl>,
#> #   test_specificity_mean <dbl>, test_specificity_std <dbl>