Optimal survey scheme

Find the optimal survey scheme that maximizes return on investment using value of information analyses. This function uses the exact method for calculating the expected value of the decision given a survey scheme.

optimal_survey_scheme(
  site_data,
  feature_data,
  site_detection_columns,
  site_n_surveys_columns,
  site_probability_columns,
  site_management_cost_column,
  site_survey_cost_column,
  feature_survey_column,
  feature_survey_sensitivity_column,
  feature_survey_specificity_column,
  feature_model_sensitivity_column,
  feature_model_specificity_column,
  feature_target_column,
  total_budget,
  survey_budget,
  site_management_locked_in_column = NULL,
  site_management_locked_out_column = NULL,
  site_survey_locked_out_column = NULL,
  prior_matrix = NULL,
  n_threads = 1,
  verbose = FALSE
)

Arguments

site_data
sf::sf() object with site data.

feature_data
base::data.frame() object with feature data.

site_detection_columns
character names of numeric columns in the argument to site_data that contain the proportion of surveys conducted within each site that detected each feature. Each column should correspond to a different feature, and contain a proportion value (between zero and one). If a site has not previously been surveyed, a value of zero should be used.

site_n_surveys_columns
character names of numeric columns in the argument to site_data that contain the total number of surveys conducted for each each feature within each site. Each column should correspond to a different feature, and contain a non-negative integer number (e.g. 0, 1, 2, 3). If a site has not previously been surveyed, a value of zero should be used.

site_probability_columns
character names of numeric columns in the argument to site_data that contain modelled probabilities of occupancy for each feature in each site. Each column should correspond to a different feature, and contain probability data (values between zero and one). No missing (NA) values are permitted in these columns.

site_management_cost_column
character name of column in the argument to site_data that contains costs for managing each site for conservation. This column should have numeric values that are equal to or greater than zero. No missing (NA) values are permitted in this column.

site_survey_cost_column
character name of column in the argument to site_data that contains costs for surveying each site. This column should have numeric values that are equal to or greater than zero. No missing (NA) values are permitted in this column.

feature_survey_column
character name of the column in the argument to feature_data that contains logical (TRUE / FALSE) values indicating if the feature will be surveyed in the planned surveys or not. Note that considering additional features will rapidly increase computational burden, and so it is only recommended to consider features that are of specific conservation interest. No missing (NA) values are permitted in this column.

feature_survey_sensitivity_column
character name of the column in the argument to feature_data that contains probability of future surveys correctly detecting a presence of each feature in a given site (i.e. the sensitivity of the survey methodology). This column should have numeric values that are between zero and one. No missing (NA) values are permitted in this column.

feature_survey_specificity_column
character name of the column in the argument to feature_data that contains probability of future surveys correctly detecting an absence of each feature in a given site (i.e. the specificity of the survey methodology). This column should have numeric values that are between zero and one. No missing (NA) values are permitted in this column.

feature_model_sensitivity_column
character name of the column in the argument to feature_data that contains probability of the initial models correctly predicting a presence of each feature in a given site (i.e. the sensitivity of the models). This column should have numeric values that are between zero and one. No missing (NA) values are permitted in this column. This should ideally be calculated using fit_xgb_occupancy_models() or fit_hglm_occupancy_models().

feature_model_specificity_column
character name of the column in the argument to feature_data that contains probability of the initial models correctly predicting an absence of each feature in a given site (i.e. the specificity of the models). This column should have numeric values that are between zero and one. No missing (NA) values are permitted in this column. This should ideally be calculated using fit_xgb_occupancy_models() or fit_hglm_occupancy_models().

feature_target_column
character name of the column in the argument to feature_data that contains the \(target\) values used to parametrize the conservation benefit of managing of each feature. This column should have numeric values that are equal to or greater than zero. No missing (NA) values are permitted in this column.

total_budget
numeric maximum expenditure permitted for conducting surveys and managing sites for conservation.

survey_budget
numeric maximum expenditure permitted for conducting surveys.

site_management_locked_in_column
character name of the column in the argument to site_data that contains logical (TRUE / FALSE) values indicating which sites should be locked in for (TRUE) being managed for conservation or (FALSE) not. No missing (NA) values are permitted in this column. This is useful if some sites have already been earmarked for conservation, or if some sites are already being managed for conservation. Defaults to NULL such that no sites are locked in.

site_management_locked_out_column
character name of the column in the argument to site_data that contains logical (TRUE / FALSE) values indicating which sites should be locked out for (TRUE) being managed for conservation or (FALSE) not. No missing (NA) values are permitted in this column. This is useful if some sites could potentially be surveyed to improve model predictions even if they cannot be managed for conservation. Defaults to NULL such that no sites are locked out.

site_survey_locked_out_column
character name of the column in the argument to site_data that contains logical (TRUE / FALSE) values indicating which sites should be locked out (TRUE) from being selected for future surveys or (FALSE) not. No missing (NA) values are permitted in this column. This is useful if some sites will never be considered for future surveys (e.g. because they are too costly to survey, or have a low chance of containing the target species). Defaults to NULL such that no sites are locked out.

prior_matrix
numeric matrix containing the prior probability of each feature occupying each site. Rows correspond to features, and columns correspond to sites. Defaults to NULL such that prior data is calculated automatically using prior_probability_matrix().

n_threads
integer number of threads to use for computation.

verbose
logical indicating if information should be printed during processing. Defaults to FALSE.

Value

matrix of logical (TRUE/ FALSE) values indicating if a site is selected in the scheme or not. Columns correspond to sites, and rows correspond to different schemes. If there is only one optimal survey scheme then the matrix will only contain a single row. This matrix also has a numeric "ev" attribute that contains the expected value of each scheme.

Details

The optimal survey scheme is determined using a brute-force algorithm. Initially, all feasible (valid) survey schemes are identified given the survey costs and the survey budget (using feasible_survey_schemes(). Next, the expected value of each and every feasible survey scheme is computed (using evdsi()). Finally, the greatest expected value is identified, and all survey schemes that share this greatest expected value are returned. Due to the nature of this algorithm, it can take a very long time to complete.

Examples

# set seeds for reproducibility
library(RandomFields)
set.seed(123)
RFoptions(seed = 123)

# simulate data
site_data <- simulate_site_data(n_sites = 30, n_features = 2, prop = 0.1)

#> ...

feature_data <- simulate_feature_data(n_features = 2, prop = 1)
feature_data$target <- c(10, 10)

# preview simulated data
print(site_data)

#> Simple feature collection with 30 features and 11 fields
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: 0.04205953 ymin: 0.02461368 xmax: 0.9942698 ymax: 0.9630242
#> CRS:            NA
#> # A tibble: 30 x 12
#>    survey_cost management_cost    f1    f2    n1    n2      e1     e2     e3
#>          <dbl>           <dbl> <dbl> <dbl> <dbl> <dbl>   <dbl>  <dbl>  <dbl>
#>  1        7.60            65.0   0.8     1     5     5 -1.59   -0.335 -1.01 
#>  2        7.87            67.8   0       1     5     5 -1.63   -2.04  -1.92 
#>  3        7.14            64.5   1       1     2     2 -0.576   0.794 -0.516
#>  4        6.90            67.6   0       1     2     2 -1.30   -2.44  -1.41 
#>  5        5.03            68.6   1       1     4     4  0.485   0.686  1.46 
#>  6        5.24            72.4   0       0     3     3  1.39   -0.181 -0.329
#>  7        8.07            68.7   0       0     0     0 -1.06    0.193 -0.880
#>  8        3.43            69.8   0.4     1     5     5  0.0778 -0.443  1.27 
#>  9        3.62            65.0   1       1     1     1  0.270   0.835  0.497
#> 10        3.68            63.1   1       0     3     3  0.657   1.18   0.366
#> # ... with 20 more rows, and 3 more variables: p1 <dbl>, p2 <dbl>,
#> #   geometry <POINT>

print(feature_data)

#> # A tibble: 2 x 7
#>   name  survey survey_sensitiv~ survey_specific~ model_sensitivi~
#>   <chr> <lgl>             <dbl>            <dbl>            <dbl>
#> 1 f1    TRUE              0.950            0.867            0.745
#> 2 f2    TRUE              0.960            0.820            0.791
#> # ... with 2 more variables: model_specificity <dbl>, target <dbl>

# set total budget for managing sites for conservation
# (i.e. 50% of the cost of managing all sites)
total_budget <- sum(site_data$management_cost) * 0.5

# set total budget for surveying sites for conservation
# (i.e. 10% of the cost of managing all sites)
survey_budget <- sum(site_data$survey_cost) * 0.1

# find optimal survey scheme using exact method
opt_survey <- optimal_survey_scheme(
  site_data, feature_data,
  c("f1", "f2"), c("n1", "n2"), c("p1", "p2"),
  "management_cost", "survey_cost",
  "survey", "survey_sensitivity", "survey_specificity",
  "model_sensitivity", "model_specificity",
  "target", total_budget, survey_budget)

# print result
print(opt_survey)

#>       [,1]  [,2]  [,3]  [,4]  [,5]  [,6]  [,7]  [,8]  [,9] [,10] [,11] [,12]
#> [1,] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [2,] FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE
#> [3,] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [4,] FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE
#> [5,] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [6,] FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE
#> [7,] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>      [,13] [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24]
#> [1,] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [2,] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [3,] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE
#> [4,] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE
#> [5,] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [6,] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [7,] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE
#>      [,25] [,26] [,27] [,28] [,29] [,30]
#> [1,] FALSE FALSE FALSE FALSE FALSE FALSE
#> [2,] FALSE FALSE FALSE FALSE FALSE FALSE
#> [3,] FALSE FALSE FALSE FALSE FALSE FALSE
#> [4,] FALSE FALSE FALSE FALSE FALSE FALSE
#> [5,] FALSE FALSE FALSE  TRUE FALSE FALSE
#> [6,] FALSE FALSE FALSE  TRUE FALSE FALSE
#> [7,] FALSE FALSE FALSE  TRUE FALSE FALSE
#> attr(,"ev")
#> [1] 1.387088 1.387088 1.387088 1.387088 1.387088 1.387088 1.387088