Prior probability matrix

Create prior probability matrix for the value of information analysis.

prior_probability_matrix(
  site_data,
  feature_data,
  site_occupancy_columns,
  site_probability_columns,
  feature_survey_sensitivity_column,
  feature_survey_specificity_column,
  feature_model_sensitivity_column,
  feature_model_specificity_column
)

Arguments

site_data
sf object with site data.

feature_data
data.frame object with feature data.

site_occupancy_columns
character names of numeric columns in the argument to site_data that contain detections and non-detections of features at each site. Each column should correspond to a different feature, and contain binary detection/non-detection value (zeros or ones) indicating if the feature was detected in a previous survey or not. If a site has not been surveyed before, then missing (NA) values should be used. Additionally, if a feature was not looked for when surveying a specific site, then missing (NA) value should be used.

site_probability_columns
character names of numeric columns in the argument to site_data that contain modelled probabilities of occupancy for each feature in each site. Each column should correspond to a different feature, and contain probability data (values between zero and one). No missing (NA) values are permitted in these columns.

feature_survey_sensitivity_column
character name of the column in the argument to feature_data that contains probability of future surveys correctly detecting a presence of each feature in a given site (i.e. the sensitivity of the survey methodology). This column should have numeric values that are between zero and one. No missing (NA) values are permitted in this column.

feature_survey_specificity_column
character name of the column in the argument to feature_data that contains probability of future surveys correctly detecting an absence of each feature in a given site (i.e. the specificity of the survey methodology). This column should have numeric values that are between zero and one. No missing (NA) values are permitted in this column.

feature_model_sensitivity_column
character name of the column in the argument to feature_data that contains probability of the initial models correctly predicting a presence of each feature in a given site (i.e. the sensitivity of the models). This column should have numeric values that are between zero and one. No missing (NA) values are permitted in this column. This should ideally be calculated using fit_occupancy_models.

feature_model_specificity_column
character name of the column in the argument to feature_data that contains probability of the initial models correctly predicting an absence of each feature in a given site (i.e. the specificity of the models). This column should have numeric values that are between zero and one. No missing (NA) values are permitted in this column. This should ideally be calculated using fit_occupancy_models.

Value

matrix object containing the prior probabilities of each feature occupying each site. Each row corresponds to a different feature and each column corresponds to a different site.

Details

The prior matrix is constructed using a combination of previous survey results and modelled predictions for sites that have not been surveyed. Let \(I\) denote the set of features (indexed by \(i\)) and let \(J\) denote the set of sites (indexed by \(j\)). Let \(D_j\) indicate which sites have already been surveyed (using zeros and ones). Let \(S_i\) and \(N_i\) denote sensitivity and specificity of the survey method for features \(i \in I\) (respectively). Also let \(H_{ij}\) denote the presence or absence of each feature in each site if they have been surveyed (using zeros or ones), with unsurveyed sites denoted with -1 values Let \({H'}_{ij}\) indicate the modelled probability of each feature occupying each site. Finally, let \({S'}_i\) and \({N'}_i\) denote sensitivity and specificity of the model for feature \(i \in I\) (respectively). Given such data, the (\(P_{ij}\)) prior probability of feature \(i\) occupying site \(j\) is:

$$ P_{ij} = \\ S_i, \text{ if } D_j = 1, H_{ij} = 1 \space (i \text{ detected in } j) \\ 1 - N_i, \text{ else if } D_j = 1, H_{ij} = 0 \space (i \text{ not detected in } j) \\ {S'}_i \times S_i, \text{ else if } D_j = 0, {H'}_{ij} \geq 0.5 \space (j \text{ not surveyed and } i \text{ predicted present in } j \text{)} \\ 1 - ({N'}_i \times N_i), \text{ else if } D_j = 0, {H'}_{ij} \geq 0.5 \space (j \text{ not surveyed and } i \text{ predicted absent in } j \text{)} \\ $$

Note that the prior probability calculations account for the fact that the species distribution models were evaluated using the survey data. Since the species distribution models were evaluated using survey data, this means that sensitivity and specificity values of the models are conditional on uncertainty present in the survey methodology. To account for this, the prior probability of species \(i\) occurring within planning unit \(j\) when relying on the species distribution model predictions depends on the sensitivity and specificity of both the species distribution models and the survey methodology.

Examples

# set seeds for reproducibility
library(RandomFields)
set.seed(123)
RFoptions(seed = 123)

# simulate data
site_data <- simulate_site_data(n_sites = 5, n_features = 2,
                                prop = 0.5)

#> ...

feature_data <- simulate_feature_data(n_features = 2, prop = 1)

# preview simulated data
print(site_data)

#> Simple feature collection with 5 features and 9 fields
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: 0.2875775 ymin: 0.0455565 xmax: 0.9404673 ymax: 0.892419
#> CRS:            NA
#> # A tibble: 5 x 10
#>   survey_cost management_cost     e1     e2     e3    f1    f2    p1     p2
#>         <dbl>           <dbl>  <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl>  <dbl>
#> 1        8.13            73.9 -0.365  0.642  1.08      1     1 0.950 0.940 
#> 2        6.07            69.7 -0.303 -0.382 -0.838     0     0 0.229 0.362 
#> 3        3.74            62.0  1.77   1.42   1.11     NA    NA 0.789 0.0228
#> 4        6.82            71.8 -0.404 -0.771 -0.732    NA    NA 0.157 0.552 
#> 5        5.22            70.7 -0.697 -0.910 -0.614     1     1 0.201 0.798 
#> # ... with 1 more variable: geometry <POINT>

print(feature_data)

#> # A tibble: 2 x 9
#>   name  survey survey_sensitiv~ survey_specific~ model_sensitivi~
#>   <chr> <lgl>             <dbl>            <dbl>            <dbl>
#> 1 f1    TRUE              0.955            0.822            0.775
#> 2 f2    TRUE              0.961            0.829            0.723
#> # ... with 4 more variables: model_specificity <dbl>, preweight <dbl>,
#> #   postweight <dbl>, target <dbl>

# calculate prior probability matrix
prior_matrix <- prior_probability_matrix(
  site_data, feature_data, c("f1", "f2"), c("p1", "p2"),
  "survey_sensitivity", "survey_specificity", "model_sensitivity",
  "model_specificity")

# preview prior probability matrix
print(prior_matrix)

#>         [,1]      [,2]      [,3]      [,4]      [,5]
#> f1 0.9554571 0.1784514 0.7403233 0.2642984 0.9554571
#> f2 0.9612699 0.1713662 0.2718601 0.6953229 0.9612699