Prior probability matrix

Create prior probability matrix for the value of information analysis.

prior_probability_matrix(
  site_data,
  feature_data,
  site_detection_columns,
  site_n_surveys_columns,
  site_probability_columns,
  feature_survey_sensitivity_column,
  feature_survey_specificity_column,
  feature_model_sensitivity_column,
  feature_model_specificity_column
)

Arguments

site_data
sf::sf() object with site data.

feature_data
base::data.frame() object with feature data.

site_detection_columns
character names of numeric columns in the argument to site_data that contain the proportion of surveys conducted within each site that detected each feature. Each column should correspond to a different feature, and contain a proportion value (between zero and one). If a site has not previously been surveyed, a value of zero should be used.

site_n_surveys_columns
character names of numeric columns in the argument to site_data that contain the total number of surveys conducted for each each feature within each site. Each column should correspond to a different feature, and contain a non-negative integer number (e.g. 0, 1, 2, 3). If a site has not previously been surveyed, a value of zero should be used.

site_probability_columns
character names of numeric columns in the argument to site_data that contain modelled probabilities of occupancy for each feature in each site. Each column should correspond to a different feature, and contain probability data (values between zero and one). No missing (NA) values are permitted in these columns.

feature_survey_sensitivity_column
character name of the column in the argument to feature_data that contains probability of future surveys correctly detecting a presence of each feature in a given site (i.e. the sensitivity of the survey methodology). This column should have numeric values that are between zero and one. No missing (NA) values are permitted in this column.

feature_survey_specificity_column
character name of the column in the argument to feature_data that contains probability of future surveys correctly detecting an absence of each feature in a given site (i.e. the specificity of the survey methodology). This column should have numeric values that are between zero and one. No missing (NA) values are permitted in this column.

feature_model_sensitivity_column
character name of the column in the argument to feature_data that contains probability of the initial models correctly predicting a presence of each feature in a given site (i.e. the sensitivity of the models). This column should have numeric values that are between zero and one. No missing (NA) values are permitted in this column. This should ideally be calculated using fit_xgb_occupancy_models() or fit_hglm_occupancy_models().

feature_model_specificity_column
character name of the column in the argument to feature_data that contains probability of the initial models correctly predicting an absence of each feature in a given site (i.e. the specificity of the models). This column should have numeric values that are between zero and one. No missing (NA) values are permitted in this column. This should ideally be calculated using fit_xgb_occupancy_models() or fit_hglm_occupancy_models().

Value

matrix object containing the prior probabilities of each feature occupying each site. Each row corresponds to a different feature and each column corresponds to a different site.

Examples

# set seeds for reproducibility
library(RandomFields)
set.seed(123)
RFoptions(seed = 123)

# simulate data
site_data <- simulate_site_data(n_sites = 5, n_features = 2, prop = 0.5)
#> ...
feature_data <- simulate_feature_data(n_features = 2, prop = 1)

# preview simulated data
print(site_data)
#> Simple feature collection with 5 features and 11 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 0.2875775 ymin: 0.0455565 xmax: 0.9404673 ymax: 0.892419
#> CRS:           NA
#> # A tibble: 5 x 12
#>   survey_cost management_cost    f1    f2    n1    n2     e1     e2     e3    p1
#>         <dbl>           <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl>  <dbl>  <dbl> <dbl>
#> 1        8.13            73.9   1       1     5     5 -0.365  0.642  1.08  0.991
#> 2        6.07            69.7   0.2     0     5     5 -0.303 -0.382 -0.838 0.126
#> 3        3.74            62.0   0       0     0     0  1.77   1.42   1.11  0.892
#> 4        6.82            71.8   0       0     2     2 -0.404 -0.771 -0.732 0.064
#> 5        5.22            70.7   0       0     0     0 -0.697 -0.910 -0.614 0.099
#> # … with 2 more variables: p2 <dbl>, geometry <POINT>
print(feature_data)
#> # A tibble: 2 x 7
#>   name  survey survey_sensitivity survey_specificity model_sensitivity
#>   <chr> <lgl>               <dbl>              <dbl>             <dbl>
#> 1 f1    TRUE                0.965              0.821             0.701
#> 2 f2    TRUE                0.960              0.814             0.701
#> # … with 2 more variables: model_specificity <dbl>, target <dbl>

# calculate prior probability matrix
prior_matrix <- prior_probability_matrix(
  site_data, feature_data,
  c("f1", "f2"), c("n1", "n2"), c("p1", "p2"),
  "survey_sensitivity", "survey_specificity",
  "model_sensitivity", "model_specificity")

# preview prior probability matrix
print(prior_matrix)
#>         [,1]         [,2]      [,3]        [,4]      [,5]
#> f1 0.9997838 1.724126e-05 0.8108543 0.001782446 0.2636165
#> f2 0.9997304 2.757199e-07 0.2523413 0.002372225 0.8597336