Simulate feature data

Simulate feature data for developing simulated survey schemes.

simulate_feature_data(n_features, proportion_of_survey_features)

Arguments

n_features
integer number of features.

proportion_of_survey_features
numeric proportion of features that will be examined in the new surveys. Values must be between zero and one.

Value

tibble object. It contains the following data:

name

character name of each feature.

survey

logical (TRUE / FALSE) values indicating if each feature should be examined in surveys or not.

survey_sensitivity

numeric sensitivity (true positive rate) of the survey methodology for each features.

survey_specificity

numeric specificity (true negative rate) of the survey methodology for each features.

model_sensitivity

numeric specificity (true positive rate) of the occupancy models for each features.

alpha

numeric values used to parametrize the conservation benefit of managing of each feature.

gamma

numeric values used to parametrize the conservation benefit of managing of each feature.

See also

Examples

# set seed for reproducibility
set.seed(123)

# simulate data
d <- simulate_feature_data(n_features = 5,
                           proportion_of_survey_features = 0.5)
# print data
print(d)

#> # A tibble: 5 x 8
#>   name  survey survey_sensitiv~ survey_specific~ model_sensitivi~
#>   <chr> <lgl>             <dbl>            <dbl>            <dbl>
#> 1 f1    FALSE             0.952            0.896            0.790
#> 2 f2    TRUE              0.971            0.845            0.725
#> 3 f3    TRUE              0.986            0.868            0.704
#> 4 f4    FALSE             0.972            0.857            0.733
#> 5 f5    TRUE              0.968            0.810            0.795
#> # ... with 3 more variables: model_specificity <dbl>, alpha <dbl>, gamma <dbl>